### 2024-02-23 #225 @editorialbot
[PRE REVIEW]: WeibullR.learnr: An Interactive Introduction to Life Data Analysis
**Submitting author:** @paulgovan (<a href="http://orcid.org/0000-0002-1821-8492">Paul Govan</a>)
**Repository:** https://github.com/paulgovan/WeibullR.learnr
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
