### 2024-02-23 #225 @labarba
@paulgovan 👋
Let me start with an apology: I am the editor-in-chief and I have neglected JOSE for several months due to plain overwhelm. I have cut down on other service and am back here with intention of reviving our little alt-journal. Thank you for your submission! Given the time that has passed, do you have updates to the paper?
