+++ pkgdown.yml
@@ -1,8 +1,8 @@
-pandoc: 2.19.2
-pkgdown: 2.0.7
+pandoc: 3.1.11
+pkgdown: 2.0.9
 pkgdown_sha: ~
 articles: {}
-last_built: 2023-07-17T19:21Z
+last_built: 2024-06-17T01:40Z
 urls:
   reference: https://paulgovan.github.io/WeibullR.learnr/reference
   article: https://paulgovan.github.io/WeibullR.learnr/articles
