+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.0.9
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-06-17T01:40Z
+last_built: 2024-07-02T16:05Z
 urls:
   reference: https://paulgovan.github.io/WeibullR.learnr/reference
   article: https://paulgovan.github.io/WeibullR.learnr/articles
