+++ .github/workflows/r.yml
@@ -32,7 +32,7 @@ jobs:
           r-version: ${{ matrix.r-version }}
       - name: Install dependencies
         run: |
-          install.packages(c("remotes", "rcmdcheck"))
+          install.packages(c("remotes", "rcmdcheck", "learnr"))
           remotes::install_deps(dependencies = TRUE)
         shell: Rscript {0}
       - name: Check
