+++ DESCRIPTION
@@ -2,8 +2,8 @@ Package: WeibullR.learnr
 Type: Package
 Title: An Interactive Introduction to Life Data Analysis
 Version: 0.1.2
-URL: https://paulgovan.github.io/WeibullR.learnr/
-BugReports: https://github.com/paulgovan/weibullr.learnr/issues
+URL: https://paulgovan.github.io/WeibullR.learnr/, https://github.com/paulgovan/WeibullR.learnr
+BugReports: https://github.com/paulgovan/WeibullR.learnr/issues
 Authors@R: person("Paul", "Govan", email = "paul.govan2@gmail.com", role = c("aut", "cre", "cph"), comment = c(ORCID = "0000-0002-1821-8492"))
 Year: 2023
 Description: An interactive introduction to Life Data Analysis that depends on 'WeibullR' 
