+++ cran-comments.md
@@ -1,13 +1,16 @@
-# Test environments
-* local OS X install, R 4.2.3
-* win-builder (devel and release)
-
 ## R CMD check results
 
 0 errors | 0 warnings | 1 note
 
-* This is a new release.
+* Maintainer: ‘Paul Govan <paul.govan2@gmail.com>’
+  
+  New maintainer:
+    Paul Govan <paul.govan2@gmail.com>
+  Old maintainer(s):
+    Paul Govan <pgovan1@aggienetwork.com>
+    
+The old email address is inactive.
 
-## Downstream dependencies
-There are currently no downstream dependencies for this package.
+## Reverse Dependencies
 
+There are currently no reverse dependencies.
