+++ DESCRIPTION
@@ -5,7 +5,6 @@ Version: 0.1.2
 URL: https://paulgovan.github.io/WeibullR.learnr/, https://github.com/paulgovan/WeibullR.learnr
 BugReports: https://github.com/paulgovan/WeibullR.learnr/issues
 Authors@R: person("Paul", "Govan", email = "paul.govan2@gmail.com", role = c("aut", "cre", "cph"), comment = c(ORCID = "0000-0002-1821-8492"))
-Year: 2023
 Description: An interactive introduction to Life Data Analysis that depends on 'WeibullR' 
   by David Silkworth and Jurgen Symynck (2022) <https://CRAN.R-project.org/package=WeibullR>, 
   a R package for Weibull Analysis, and 'learnr' by Garrick Aden-Buie et al. (2023) 
