+++ inst/paper/paper.bib
@@ -36,7 +36,7 @@
   author = {Paul Govan},
   year = {2023},
   note = {R package version 0.1.1},
-  url = {https://paulgovan.github.io/WeibullR.learnr/},
+  url = {https://CRAN.R-project.org/package=WeibullR.learnr},
 }
 
 @Manual{WeibullRplotly,
@@ -52,7 +52,7 @@
     author = {Paul Govan},
     year = {2023},
     note = {R package version 0.1.0},
-    url = {https://paulgovan.github.io/WeibullR.shiny/},
+    url = {https://CRAN.R-project.org/package=WeibullR.shiny},
   }
 
 @book{Weibull,
