+++ .github/workflows/draft-pdf.yml
@@ -12,7 +12,7 @@ jobs:
         with:
           journal: jose
           # This should be the path to the paper within your repo.
-          paper-path: paper.md
+          paper-path: inst/paper/paper.md
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
@@ -20,4 +20,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: paper.pdf
+          path: inst/paper/paper.pdf
