+++ inst/tutorials/RAMRlearnr/RAMRlearnr.Rmd
@@ -12,21 +12,25 @@ library(learnr)
 library(WeibullR.learnr)
 ```
 
-## Reliability, Availability and Maintainability
+## Introduction
 
-Reliability, Availability, and Maintainability may seem like similar terms, but they actually represent different concepts. 
+Welcome to the tutorial on **Reliability, Availability, and Maintainability (RAM)**! In this tutorial, you will learn about these fundamental concepts and their applications.
 
-**Reliability** is the ability of an item to not break down over time. For example, a reliable car is one that does not break down very often.
+## What is RAM?
 
-**Availability** is the ability of an item to be available for use. An available car is one waiting patiently in the garage and not the shop. 
+The terms Reliability, Availability, and Maintainability may sound similar, but they actually represent different concepts.
 
-**Maintainability** is the ability of an item to be maintained efficiently. A maintainable car is quick to fix if it has a problem.
+**Reliability**: The ability of an item to perform its intended function without failure over a specified period.
 
-That being said, all 3 concepts are inter-related. For example, a car that is not very reliable is probably not available very often. A car that is also difficult to maintain is probably available even less. 
+**Availability**: The proportion of time an item is in a functioning condition.
+
+**Maintainability**: The ease and speed with which an item can be restored to operational status after a failure.
+
+That being said, all 3 terms are inter-related. A car that is not very reliable is probably not available very often. A car that is also difficult to maintain is probably available even less. 
 
 ## Learning Objectives
 
-This learning module provides a quick reference for basic Reliability, Availability, and Maintainability concepts. After completing this module, you should have a basic understanding of and a chance to apply the following concepts:
+This learning module provides a quick reference for basic RAM concepts. After completing this module, you should have a basic understanding of and a chance to apply the following concepts:
 
 * Reliability
 * Availability
@@ -236,10 +240,13 @@ The $B_n$ life for the Exponential distribution can be estimated with the base f
 qexp(0.1, 0.1)
 ```
 
+## Summary
 
-### References
+In this tutorial, we've introduced the concepts of Reliability, Availability, and Maintainability. You have learned to calculate reliability using different assumptions and explored the importance of these concepts in system design.
+
+## References
 
 - Abernethy, R.B. (2004) The New Weibull Handbook. Fifth Edition.
 
--   Aden-Buie G, Schloerke B, Allaire J, Rossell Hayes A (2023). learnr: Interactive Tutorials for R. https://rstudio.github.io/learnr/, https://github.com/rstudio/learnr.
+- Aden-Buie G, Schloerke B, Allaire J, Rossell Hayes A (2023). learnr: Interactive Tutorials for R. https://rstudio.github.io/learnr/, https://github.com/rstudio/learnr.
 
