+++ R/WeibullR.learnr.R
@@ -1,15 +1,14 @@
-#' An Interactive Introduction to Life Data Analysis
+#' An Interactive Introduction to Reliability Testing
 #'
-#' WeibullR.learnr is an interactive introduction to Life Data Analysis.
-#' @import learnr
-#' @import WeibullR
+#' TestR.learnr is an interactive introduction to Reliability Testing.
+#' @import learnr ReliaGrowR WeibullR.ALT
 #' @export
 #' @seealso \url{https://paulgovan.github.io/WeibullR.learnr/}
 #' @return This function does not return a value.
 #' @examples
 #' if (interactive()) {
-#'   WeibullR.learnr()
+#'   TestR.learnr()
 #' }
-WeibullR.learnr <- function() {
-  learnr::run_tutorial('WeibullRlearnr', package = 'WeibullR.learnr')
+TestR.learnr <- function() {
+  learnr::run_tutorial('TestRlearnr', package = 'WeibullR.learnr')
 }
