+++ man/TestR.learnr.Rd
@@ -1,20 +1,20 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/WeibullR.learnr.R
-\name{WeibullR.learnr}
-\alias{WeibullR.learnr}
-\title{An Interactive Introduction to Life Data Analysis}
+\name{TestR.learnr}
+\alias{TestR.learnr}
+\title{An Interactive Introduction to Reliability Testing}
 \usage{
-WeibullR.learnr()
+TestR.learnr()
 }
 \value{
 This function does not return a value.
 }
 \description{
-WeibullR.learnr is an interactive introduction to Life Data Analysis.
+TestR.learnr is an interactive introduction to Reliability Testing.
 }
 \examples{
 if (interactive()) {
-  WeibullR.learnr()
+  TestR.learnr()
 }
 }
 \seealso{
