+++ NAMESPACE
@@ -1,12 +1,14 @@
 # Generated by roxygen2: do not edit by hand
 
 export(RAMR.learnr)
-export(WeibullR.learnr)
+export(TestR.learnr)
 export(avail)
 export(fr)
 export(mtbf)
 export(mttf)
 export(rel)
 export(serv)
+import(ReliaGrowR)
 import(WeibullR)
+import(WeibullR.ALT)
 import(learnr)
