+++ _pkgdown.yml
@@ -10,3 +10,6 @@ tutorials:
 - name: RAMRlearnr
   title: Reliability, Availability, and Maintainability (RAM)
   url: https://paulgovan.shinyapps.io/ramrlearnr/
+- name: TestRlearnr
+  title: An Introduction to Reliability Testing
+  url: https://govan.shinyapps.io/testrlearnr/
