+++ news.md
@@ -1,3 +1,8 @@
+# WeibullR.learnr 0.2
+
+* Now with a new Reliability Testing tutorial called `TestR.learnr()` that covers Reliability Growth Analysis (RGA) and Accelerated Life Testing (ALT)
+* Other minor improvements and bug fixes
+
 # WeibullR.learnr 0.1.3
 
 ## Minor Improvements and bug fixes
