### 2025-01-29 35c6053
Update RAM definitions
