### 2025-03-03 #225 @stats-tgeorge
Good morning @paulgovan. Do you still want your submission to be reviewed and published with JOSE? If so, I will start to look for reviews (I see your suggestion but they are in high demand).
