+++ DESCRIPTION
@@ -18,6 +18,6 @@ Imports:
 Suggests:
   WeibullR.plotly,
   WeibullR.shiny
-License: CC BY 4.0
+License: Apache License (>= 2)
 Encoding: UTF-8
 RoxygenNote: 7.3.2
