### 2025-03-05 #225 @paulgovan
Ok I updated the license type.
