+++ inst/paper/paper.md
@@ -19,19 +19,21 @@ bibliography: paper.bib
 
 # Statement of Need
 
-Life data analysis is the study of how things, from machines to people, function over time. Life data shows how long things last or how long before they need to be repaired or replaced. Although numerous learning resources exist, most reference proprietary software that is often unavailable to those hoping to get started with life data. This proprietary software is usually meant for commercial applications and may not be affordable for students or young professionals. `WeibullR.learnr` [@WeibullRlearnr] is an open source set of interactive learning modules, exercises, and functions for learning introductory life data analysis. The primary motivation for the development of `WeibullR.learnr` was to simultaneously introduce the basic concepts and also open source software for analyzing life data. The target audience for this project are beginning practitioners and university students. 
+Life data analysis is the study of how systems, from machines to people, perform over time. Life data includes lifespans, failure rates, and repair or replacement timelines. While various learning resources exist, many rely on proprietary software that can be inaccessible to students and early-career professionals due to cost constraints.
+
+`WeibullR.learnr` [@WeibullRlearnr] is an open-source collection of interactive learning modules, exercises, and functions designed for introductory life data analysis. The primary goal of this project is to introduce fundamental concepts while providing an open-source alternative for analyzing life data. The target audience includes beginner practitioners and university students.
 
 # Description
 
-`WeibullR.learnr` is written in R [@R] and is built using `WeibullR` [@WeibullR], a R package for Life Data Analysis in the tradition of Waloddi Weibull [@Weibull], and `learnr` [@learnr], a framework for building interactive learning modules in R.
+`WeibullR.learnr` is written in R [@R] and utilizes `WeibullR` [@WeibullR], an R package for Life Data Analysis based on the methodologies of Waloddi Weibull [@Weibull], and `learnr` [@learnr], a framework for building interactive learning modules in R.
 
-Currently, three primary learning modules exist. These modules can be taken in either order and can be taken separately or together. The learning modules are designed to be plug-and-play, but changes can be made by forking the software repository and modifying the fork.
+Currently, three primary learning modules are available. These modules are independent and can be completed in any order. They are designed to be plug-and-play, but users can modify them by forking the repository.
 
-`WeibullR.learnr()` provides an interactive introduction to Life Data Analysis. The learning objectives include basic Weibull analysis, data censoring including right censored data and interval censored data, different types of Weibull models including the 2P Weibull, 3P Weibull, and Weibayes model, parameter estimation methods Median Rank Regression (MRR) and Maximum Likelihood Estimation (MLE), and different plotting methods such as Probability Plots and Contour Plots. The estimation duration for this module is approximately 2 hours. 
+`WeibullR.learnr()` provides an interactive introduction to Life Data Analysis. The learning objectives include basic Weibull analysis, censoring techniques such as right and interval censored data, different types of Weibull models including the 2P Weibull, 3P Weibull, and Weibayes model, parameter estimation methods Median Rank Regression (MRR) and Maximum Likelihood Estimation (MLE), and data visualization methods such as Probability Plots and Contour Plots. The estimated duration for this module is approximately 2 hours. 
 
-`RAMR.learnr()` is a quick reference for common Reliability, Availability, and Maintainability (RAM) concepts. The learning objectives include the basic concepts and application of Reliability, Availability, Mean Time to Repair (MTTR), Mean Time to Failure (MTTF), Mean Time Between Failures (MTBF), Failure Rate, Probability of Failure, and $B_n$ or $L_n$ life. The esimated duration of this module is about 1 hour.
[…]
