### 2025-03-06 #225 @stats-tgeorge
Good morning, Kelly (@kbodwin)! I hope you are doing well. It may have been a while, but you previously expressed interest in helping review JOSE (Journal of Open Source Education) submissions. These are open reviews of educational materials. Would you be willing to review this submission? Thank you!
