### 2025-07-16 3355f52
CRAN checks
