+++ cran-comments.md
@@ -4,4 +4,7 @@
 
 ## Reverse Dependencies
 
-There are currently no reverse dependencies.
+We checked 2 reverse dependencies, comparing R CMD check results across CRAN and dev versions of this package.
+
+ * We saw 0 new problems
+ * We failed to check 0 packages
