+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: WeibullR.learnr
 Type: Package
 Title: An Interactive Introduction to Life Data Analysis
-Version: 0.2
+Version: 0.2.1
 URL: https://paulgovan.github.io/WeibullR.learnr/, https://github.com/paulgovan/WeibullR.learnr
 BugReports: https://github.com/paulgovan/WeibullR.learnr/issues
 Authors@R: person("Paul", "Govan", email = "paul.govan2@gmail.com", role = c("aut", "cre", "cph"), comment = c(ORCID = "0000-0002-1821-8492"))
