### 2025-08-11 #302 @editorialbot
[REVIEW]: ReliaLearnR: Learning Modules for Reliability Analysis
**Submitting author:** @paulgovan (<a href="http://orcid.org/0000-0002-1821-8492">Paul Govan</a>)
**Repository:** https://github.com/paulgovan/WeibullR.learnr
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.1
