+++ inst/paper/paper.bib
@@ -4,6 +4,7 @@
   year = {2023},
   note = {R package version 0.11.3},
   url = {https://CRAN.R-project.org/package=learnr},
+  doi = {10.32614/CRAN.package.learnr}
 }
 
  @Manual{R,
@@ -13,6 +14,7 @@
     address = {Vienna, Austria},
     year = {2023},
     url = {https://www.R-project.org/},
+    doi = {10.32614/r.manuals}
   }
 
 @Manual{shiny,
@@ -21,6 +23,7 @@
     year = {2022},
     note = {R package version 1.7.4},
     url = {https://CRAN.R-project.org/package=shiny},
+    doi = {10.32614/CRAN.package.shiny}
   }
 
 @Manual{WeibullR,
@@ -29,6 +32,7 @@
   year = {2022},
   note = {R package version 1.2.1},
   url = {https://CRAN.R-project.org/package=WeibullR},
+  doi = {10.32614/CRAN.package.WeibullR}
 }
 
 @Manual{WeibullRlearnr,
@@ -37,6 +41,7 @@
   year = {2023},
   note = {R package version 0.1.1},
   url = {https://CRAN.R-project.org/package=WeibullR.learnr},
+  doi = {10.32614/CRAN.package.WeibullR.learnr}
 }
 
 @Manual{WeibullRplotly,
@@ -45,6 +50,7 @@
   year = {2023},
   note = {R package version 0.1.5},
   url = {https://CRAN.R-project.org/package=WeibullR.plotly},
+  doi = {10.32614/CRAN.package.WeibullR.plotly}
 }
 
 @Manual{WeibullRshiny,
@@ -53,6 +59,7 @@
     year = {2023},
     note = {R package version 0.1.0},
     url = {https://CRAN.R-project.org/package=WeibullR.shiny},
+    doi = {10.32614/CRAN.package.WeibullR.shiny}
   }
 
 @book{Weibull,
