### 2025-08-12 #302 @paulgovan
Thanks @stats-tgeorge. I added DOIs where available.
