### 2025-08-28 paulgovan/WeibullR.learnr#6 @lwjohnst86
Update date in `paper.md` (it's put as 2023)
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302
