### 2025-08-28 paulgovan/ReliaLearnR#7 @lwjohnst86
Confusing package name, when there are other tutorials within (e.g. RAMR)
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

What's the reason for naming it `WeibullR.learnr` when there are three tutorials that covert three different things? WeibullR, RAMR, and TestR?
