### 2025-08-28 paulgovan/WeibullR.learnr#8 @lwjohnst86
Target audience of tutorials ("beginners") seems disconnected from actual content
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

The very first tutorial (WeibullR) says it aims to be covered in 2 hrs, but also says it will cover MLE, Weibull (2P and 3P), MMR, and plotting. Just one of these concepts might be appropriate for a beginner tutorial, and definitely more than 2 hours. So, who is the actual audience here? Intermediate R users who have some level of statistical knowledge? Or is the aim simply to inform or draw awareness, but not have them actually learn anything?
