### 2025-08-28 paulgovan/ReliaLearnR#9 @lwjohnst86
Package could be split into three separate repos for each tutorial
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

The package contents are semantically and conceptually quite large. From a learner, user (teacher), and maintainer point of view, having three separate repos for the three tutorials makes it much easier to "digest". As it stands, the package is trying to do too much with too little.
