### 2025-08-28 paulgovan/ReliaLearnR#10 @lwjohnst86
Purpose of the helper functions?
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

It isn't clear to me what the purpose of the helper functions are (e.g. `serv()` or `mttf()`) as they aren't used anywhere in the tutorials nor are they particularly complex (mostly just doing arithmetic). Are they actually needed for the tutorials and if so, please use them then ☺️
