### 2025-08-28 paulgovan/WeibullR.learnr#11 @lwjohnst86
There is no `WeibullR.learnr()` function
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

There are only the functions for the RAM and TestR tutorials.
