### 2025-08-28 paulgovan/WeibullR.learnr#12 @lwjohnst86
Not sure what the purpose is of the `rsconnect/` folders, they don't seem to be used
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

I can't debug what the purpose of having them is. Are they needed?
