### 2025-08-28 paulgovan/ReliaLearnR#20 @lwjohnst86
Learning objectives for WeibullR tutorial are not clear
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

Learning objectives should be very clear and precise. There are many "frameworks" for writing them, such as Bloom's Taxonomy or Significant Learning taxonomy. They should describe what the learner should concretely gain by the end.

In general, words like "understand" are too vague to describe what learning will actually take place. How would you define "understand"? And how will you know when "understanding" has happened?
