### 2025-08-28 paulgovan/WeibullR.learnr#21 @lwjohnst86
Individual segments need more explanation about new language introduced
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

Whenever a new or unfamiliar new gets introduced, it needs to be explained. E.g. what does a Weibull distribution mean? The first two sections are "The Weibull distribution" and "Why Weibull". What about "what is Weibull?". This happens throughout, where a new term is used but not explained. (Keeping in mind, I'm looking at this from a beginner. If you assume a beginner, you need to assume limited prior knowledge).
