### 2025-08-28 paulgovan/WeibullR.learnr#22 @lwjohnst86
Any code always needs to be explained and described
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

Throughout the tutorial, code is shown, but never described. Whenever something is introduced, it needs to be explained or described.
