### 2025-08-28 paulgovan/ReliaLearnR#23 @lwjohnst86
Exercises need to be preceded by actual practical usage
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

Exercises are used to *test* a learners knowledge. But if there are no opportunities to gain that knowledge in the first place, and to know what to actually do during the exercises, then learning doesn't really happen, only confusion.

E.g. for the first exercise, how am I supposed to know how to answer that question if I've never been shown how to answer similar looking questions?

Lessons should generally follow the "introduce and explain a topic -> walk through an example of using it or applying it -> have the learner independently do the exercise to reinforce learning".
