### 2025-08-28 #302 @lwjohnst86
@paulgovan and @stats-tgeorge there are a lot of things that I've added (as PRs and as Issues) that I think represent a substantial amount of work to correct and address. I have so far only reviewed half of the WeibullR tutorial (not even the other two ones). Before I put anymore time into this review, I'll let @paulgovan address them. Sound good? ☺️
