### 2025-08-29 #302 @AgnesDuah
@paulgovan and @stats-tgeorge, I’ve opened an issue with details here: https://github.com/openjournals/jose-reviews/issues/305#issue-3365289969
