### 2025-08-29 #302 @paulgovan
> [@paulgovan](https://github.com/paulgovan) and [@stats-tgeorge](https://github.com/stats-tgeorge), I’ve opened an issue with details here: [#304 (comment)](https://github.com/openjournals/jose-reviews/issues/304#issue-3365270206)

Hi @AgnesDuah, please open this issue [here](https://github.com/paulgovan/WeibullR.learnr/issues)
