### 2025-08-29 paulgovan/ReliaLearnR#24 @AgnesDuah
Demonstrate how modules can be applied in teaching or training settings
The documentation provides helpful installation instructions, examples, and references to related resources. However, the paper does not demonstrate how the modules have been utilized in classroom or training settings. Including such examples would strengthen its practical relevance.
