### 2025-08-29 paulgovan/WeibullR.learnr#25 @AgnesDuah
Add code comments and graph explanations for clarity
Currently, all codes are presented without comments, which may make it difficult for beginners to follow along. Adding brief code annotations and providing a little more explanation of the graphs and outputs could make the materials more accessible and supportive for new learners.
