### 2025-08-29 #302 @AgnesDuah
@paulgovan ,I’ve opened the issue again here: https://github.com/paulgovan/WeibullR.learnr/issues/25#issue-3366909008
