### 2025-09-01 paulgovan/WeibullR.learnr#7 @paulgovan
WeibullR.learnr was the first tutorial in this package, which grew over time to cover other topics.
