### 2025-09-01 paulgovan/WeibullR.learnr#8 @paulgovan
These tutorials are adapted from a reliability program originally developed at a large technology company, where each module typically took 1–2 hours to complete. Feedback is welcome on whether the pacing aligns with user experience.

The tutorials are designed for practitioners who are new to reliability engineering—those who may have little or no prior experience. The focus is on practical, introductory material rather than deep theory, with the goal of helping users to quickly build momentum and confidence while getting started in reliability.
