### 2025-09-01 paulgovan/WeibullR.learnr#10 @paulgovan
Most of the helper functions are covered in the RAM tutorial and are provided as a quick reference for how these quantities are calculated. I removed the functions that are not included in the tutorial.
