### 2025-09-01 d992bb5
Remove serv function
