### 2025-09-01 a5d6cf4
Add WeibullR.learnr function
