+++ man/WeibullR.learnr.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/WeibullR.learnr.R
+\name{WeibullR.learnr}
+\alias{WeibullR.learnr}
+\title{An Interactive Introduction to Life Data Analysis}
+\usage{
+WeibullR.learnr()
+}
+\value{
+This function does not return a value.
+}
+\description{
+WeibullR.learnr is an interactive introduction to life data analysis.
+}
+\examples{
+if (interactive()) {
+  WeibullR.learnr()
+}
+}
+\seealso{
+\url{https://paulgovan.github.io/WeibullR.learnr/}
+}
