+++ man/TestR.learnr.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/WeibullR.learnr.R
+% Please edit documentation in R/TestR.learnr.R
 \name{TestR.learnr}
 \alias{TestR.learnr}
 \title{An Interactive Introduction to Reliability Testing}
