### 2025-09-01 d3efae2
Update learning objectives
