### 2025-09-02 a1bf08c
Update paper.md
