+++ inst/paper/paper.md
@@ -1,16 +1,16 @@
 ---
-title: 'WeibullR.learnr: An Interactive Introduction to Life Data Analysis'
+title: 'WeibullR.learnr: An Interactive Introduction to Reliability Engineering'
 tags:
   - R
   - life data analysis
   - Weibull analysis
-  - reliability
+  - reliability engineering
 authors:
   - name: Paul B. Govan
     orcid: 0000-0002-1821-8492
     affiliation: 1
 affiliations:
- - name: Technical Leader, Reliability and Data Analytics, GE Renewable Energy
+ - name: Senior Reliability Systems Engineer, Safety and Reliability, GE Aerospace
    index: 1
 date: 31 August 2025
 bibliography: paper.bib
@@ -19,27 +19,60 @@ bibliography: paper.bib
 
 # Statement of Need
 
-Life data analysis is the study of how systems, from machines to people, perform over time. Life data includes lifespans, failure rates, and repair or replacement timelines. While various learning resources exist, many rely on proprietary software that can be inaccessible to students and early-career professionals due to cost constraints.
+Life data analysis is the study of how systems, from machines to people, perform 
+over time. Life data includes lifespans, failure rates, and repair or replacement
+timelines. While various learning resources exist, many rely on proprietary software
+that can be inaccessible to students and early-career professionals due to cost constraints.
 
-`WeibullR.learnr` [@WeibullRlearnr] is an open-source collection of interactive learning modules, exercises, and functions designed for introductory life data analysis. The primary goal of this project is to introduce fundamental concepts while providing an open-source alternative for analyzing life data. The target audience includes beginner practitioners and university students.
+`WeibullR.learnr` [@WeibullRlearnr] is an open-source collection of interactive 
+learning modules, exercises, and functions designed for introductory life data analysis. 
+The primary goal of this project is to introduce fundamental concepts while providing 
+an open-source alternative for analyzing life data. The target audience includes 
+beginner practitioners and university students.
 
 # Description
 
-`WeibullR.learnr` is written in R [@R] and utilizes `WeibullR` [@WeibullR], an R package for Life Data Analysis based on the methodologies of Waloddi Weibull [@Weibull], and `learnr` [@learnr], a framework for building interactive learning modules in R.
+`WeibullR.learnr` is written in R [@R] and utilizes `WeibullR` [@WeibullR], an R 
+package for Life Data Analysis based on the methodologies of Waloddi Weibull [@Weibull],
+and `learnr` [@learnr], a framework for building interactive learning modules in R.
 
-Currently, three primary learning modules are available. These modules are independent and can be completed in any order. They are designed to be plug-and-play, but users can modify them by forking the repository.
+Currently, three primary learning modules are available. These modules are independent
+and can be completed in any order. They are designed to be plug-and-play, but users 
+can modify them by forking the repository.
 
-`WeibullR.learnr()` provides an interactive introduction to Life Data Analysis. The learning objectives include basic Weibull analysis, censoring techniques such as right and interval censored data, different types of Weibull models including the 2P Weibull, 3P Weibull, and Weibayes model, parameter estimation methods Median Rank Regression (MRR) and Maximum Likelihood Estimation (MLE), and data visualization methods such as Probability Plots and Contour Plots. The estimated duration for this module is approximately 2 hours. 
+`WeibullR.learnr()` provides an interactive introduction to Life Data Analysis. 
+The learning objectives include basic Weibull analysis, censoring techniques such 
+as right and interval censored data, different types of Weibull models including 
+the 2P Weibull, 3P Weibull, and Weibayes model, parameter estimation methods Median
+Rank Regression (MRR) and Maximum Likelihood Estimation (MLE), and data visualization
[…]
