### 2025-09-02 paulgovan/WeibullR.learnr#24 @paulgovan
I expanded on the background and inspiration for the modules in the second to last paragraph of the paper. See Commit
