### 2025-09-06 #302 @dbaidoo
Hi all — thanks for the assignment.
COI/CoC: I’ve read the JOSE COI policy and Code of Conduct; no conflicts to declare.

Plan:
- Environment: macOS 14 / Python 3.11 (fresh conda env).
- I’ll test installation, run the examples, and check learning objectives & docs.
- I’ll open focused issues in the submission repo (linked above) and reference them here.

ETA: First pass by Sept 10, full review by Sept 20.
