### 2025-09-06 #302 @dbaidoo
git clone https://github.com/paulgovan/WeibullR.learnr.git
cd WeibullR.learnr
