### 2025-09-06 #302 @dbaidoo
### Summary
I have completed my review of **WeibullR.learnr**. The repository cloned and installed without issues, the modules (`WeibullR.learnr()`, `RAMR.learnr()`, `TestR.learnr()`) ran smoothly, and the helper functions returned expected results. The documentation is clear, well-structured, and aligned with the stated learning objectives. The paper is concise, provides sufficient background, and highlights the educational contribution effectively.

### Code Verification
Running selected helper functions produced correct outputs:

```r
# Mean time to failure
mttf(5, 1000)
[1] 200

# Reliability for a period
rel(100, 1000)
[1] 0.9

# Availability & failure rate quick checks
avail(900, 1000)
[1] 0.1

fr(5, 1000)
[1] 0.005
