### 2025-09-06 #302 @dbaidoo
### Strengths
- Clear and reproducible installation and usage instructions.
- Learning objectives and estimated time commitments are well documented.
- Interactive modules are engaging and run without errors.
- Helper functions add useful applied value for learners.
- Paper.md is well written and appropriately references related work.

### Recommendation
**Accept pending archive DOI update.**
