### 2025-09-06 paulgovan/ReliaLearnR#26 @AgnesDuah
Add README note about updating dependencies during installation
## Summary
When installing `WeibullR.learnr` from GitHub, I was prompted to update several dependencies
(e.g., Rcpp, rlang, fs, glue). Installation worked after updating, but the prompt could be confusing
for new users.

## Why this matters
Since this package is designed as an educational tool, many learners may be new to R.
If they see dependency update prompts without guidance, they might think the installation failed.
Documenting this step would help them complete the setup more smoothly.

## Suggested improvement
Add a short note in the README, for example:

"Note: You may need to update your CRAN packages before installing `WeibullR.learnr`.
You can do this by running update.packages(), or when prompted during installation type 1 (All) to update."

This small addition will improve the user experience, especially for beginners.
