### 2025-09-06 #302 @AgnesDuah
@paulgovan,I’ve opened an issue here: https://github.com/paulgovan/WeibullR.learnr/issues/26#issue-3390750575
