### 2025-09-08 paulgovan/WeibullR.learnr#27 @AgnesDuah
Interpreting β for Infant Mortality Failure Modes
In Weibull reliability testing, β < 1 (e.g., 0.5) indicates infant mortality. Still, since β must be estimated from failure data, the question should clarify whether it is testing theory (0.5) or practice (“Unable to be determined”).
