### 2025-09-08 #302 @AgnesDuah
@paulgovan,I’ve opened an issue here:https://github.com/paulgovan/WeibullR.learnr/issues/27#issue-3392206224
