### 2025-09-08 #302 @AgnesDuah
I’ve completed my review of WeibullR.learnr. I’ve opened several issues to make the feedback easier to track, covering points such as clarifying learning objectives, adding explanations before exercises, improving code comments and graph interpretations, and including installation/dependency notes. Please see issues #24–#27 for the details. Thanks again for sharing this package. I think these changes will make the tutorials even stronger for learners.
