+++ inst/tutorials/WeibullRlearnr/WeibullRlearnr.Rmd
@@ -47,10 +47,10 @@ The Weibull is a **flexible** distribution that can fit many different types of 
 
 Take the following normal distribution with a mean of 5 and standard deviation of 1.5. 
 
-```{r echo=TRUE}
-x <- seq(0, 10, by = .1)
-y <- dnorm(x, mean = 5, sd = 1.5)
-plot(x,y, type = "l", col = "blue", xlab = "Variable", ylab = "Probability")
+```{r echo=FALSE}
+x <- seq(0, 10, by = .1) # Create a sequence of x values
+y <- dnorm(x, mean = 5, sd = 1.5) # Calculate the normal density function
+plot(x,y, type = "l", col = "blue", xlab = "Variable", ylab = "Probability") # Plot the normal distribution
 ```
 
 ### The Weibull vs Normal
@@ -58,11 +58,12 @@ plot(x,y, type = "l", col = "blue", xlab = "Variable", ylab = "Probability")
 The Weibull can *mimic* the normal and many others. Adjust the scale parameter in the slider below until the Weibull looks similar to the Normal above.
 
 ```{r, echo=FALSE}
-sliderInput("eta", "Eta:", min = 1, max = 10, value = 1, step = 0.1)
-plotOutput("distPlot")
+sliderInput("eta", "Eta:", min = 1, max = 10, value = 1, step = 0.1) # Slider for eta
+plotOutput("distPlot") # Plot output
 ```
 
 ```{r, context="server"}
+# Plot the Weibull distribution
 output$distPlot <- renderPlot({
   x <- seq(0, 10, by = .1)
   y <- dweibull(x, shape = 3.67, scale = input$eta)
@@ -118,11 +119,12 @@ Let's vary the parameters of the Weibull distribution to see these relationships
 Varying Beta changes the **shape** of the cumulative distribution function. Use the slider below to adjust the shape of the Weibull. 
 
 ```{r, echo=FALSE}
-sliderInput("beta", "Beta:", min = 0.5, max = 3, value = 1)
-plotOutput("cummPlot")
+sliderInput("beta", "Beta:", min = 0.5, max = 3, value = 1) # Slider for beta
+plotOutput("cummPlot") # Plot output
 ```
 
 ```{r, context="server"}
+# Plot the cumulative distribution function
 output$cummPlot <- renderPlot({
   x <- seq(0, 50, by = .1)
   plot(pweibull(x, shape = input$beta, scale = 10), col = "blue", type = "l", xlab = "Variable", ylab = "Cumulative Probability") 
@@ -134,11 +136,12 @@ output$cummPlot <- renderPlot({
 Varying Eta moves the weibull distribution along the x-axis. Use the slider below to adjust the scale of the Weibull. 
 
 ```{r, echo=FALSE}
-sliderInput("eta2", "Eta:", min = 1, max = 20, value = 1)
-plotOutput("cummPlot2")
+sliderInput("eta2", "Eta:", min = 1, max = 20, value = 1) # Slider for eta
+plotOutput("cummPlot2") # Plot output
 ```
 
 ```{r, context="server"}
+# Plot the cumulative distribution function
 output$cummPlot2 <- renderPlot({
   x <- seq(0, 50, by = .1)
   plot(pweibull(x, shape = 1, scale = input$eta2), col = "red", type = "l", xlab = "Variable", ylab = "Cumulative Probability") 
@@ -198,12 +201,14 @@ question("What is the approximate reliability at time 50?)",
 
 Right censored data includes **suspensions** or units that have operated for a period of time without failure. 
 
-Let's add right censored data to the previous example. In addition to the 5 machines that failed, add suspensions for 3 machines that did not fail (right censored) at times 100, 45, and 10 respectively.
[…]
