### 2025-09-17 paulgovan/WeibullR.learnr#21 @paulgovan
I went back through the modules looking for any new terms that were unexplained. I believe I caught most of them, but let me know if anything else is unclear.
