### 2025-09-17 paulgovan/WeibullR.learnr#25 @paulgovan
I went back through the code and added comments where missing and more explanations of the graphs. See also https://github.com/paulgovan/WeibullR.learnr/issues/22.
