### 2025-09-17 paulgovan/WeibullR.learnr#26 @paulgovan
Good point! I added a note to the README. See fa1689c7cdcfcdb014b6a27bb622cb017df2c0f3.
