### 2025-09-17 3e36864
Update exercises
