+++ inst/tutorials/RAMRlearnr/RAMRlearnr.Rmd
@@ -47,15 +47,24 @@ Unreliability is the opposite of reliability, that is, the probability that an i
 
 $$ Unreliability = 1 - Reliability = (Failed Time/Total Time)*100 $$
 
-### Example
-
-```{r reliability, echo=FALSE}
-question("An item has run for 3 years total and was failed for 5 of those days. What is the reliability of the item?",
-  answer("90%"),
-  answer("0.5%"),
-  answer("99.5%", correct = TRUE),
-  answer("75%"),
-  random_answer_order = TRUE
+```{r quiz1, echo=FALSE}
+quiz(caption = "Exercise: Reliability",
+  question("An item has run for 3 years total and was failed for 5 of those days. What is the reliability of the item?",
+    answer("90%", message = "Not quite, try again."),
+    answer("0.5%", message = "0.5% is the unreliability, not the reliability."),
+    answer("99.5%", correct = TRUE),
+    answer("75%", message = "Not quite, try again."),
+    random_answer_order = TRUE,
+    allow_retry = TRUE
+  ),
+  question("What is the unreliability of the item?",
+    answer("0.5%", correct = TRUE),
+    answer("90%", message = "Not quite, try again."),
+    answer("99.5%", message = "99.5% is the reliability, not the unreliability."),
+    answer("75%", message = "Not quite, try again."),
+    random_answer_order = TRUE,
+    allow_retry = TRUE
+  )
 )
 ```
 
@@ -68,12 +77,23 @@ $$ Availability = (1 - (Unavailable Time/Total Time))*100 $$
 Unavailable time is different than standby time, where standby time is when the item is available for service, but is not being used. 
 
 ```{r availability, echo=FALSE}
-question("An item has run for 3 years total, was failed for 5 days, and had scheduled maintenance for 14 days. What is the availability of the item?",
-  answer("95%"),
-  answer("1.7%"),
-  answer("98.3%", correct = TRUE),
-  answer("50%"),
-  random_answer_order = TRUE
+quiz(caption = "Exercise: Availability",
+     question("What is the difference between reliability and availability?",
+       answer("Reliability includes scheduled maintenance, availability does not.", message = "Not quite, try again."),
+       answer("Availability includes scheduled maintenance, reliability does not.", correct = TRUE),
+       answer("There is no difference, they are the same.", message = "There is a slight difference, try again."),
+       answer("Reliability is always larger than availability.", message = "Not always, try again."),
+       random_answer_order = TRUE,
+       allow_retry = TRUE
+     ),
+  question("An item has run for 3 years total, was failed for 5 days, and had scheduled maintenance for 14 days. What is the availability of the item?",
+    answer("95%", message = "Not quite, try again."),
+    answer("1.7%", message = "1.7% is too low, try again."),
+    answer("98.3%", correct = TRUE),
+    answer("50%", message = "50% is too low, try again."),
+    random_answer_order = TRUE,
+  allow_retry = TRUE
+  )
 )
 ```
 
@@ -83,43 +103,65 @@ The MTTR is a measure of maintainability and is the mean time to repair an item.
 
 $$ MTTR = \sum_{i=1}^n RepairTime_i/RepairCount $$
 
-```{r mttr, echo=FALSE}
-question("1000 items run for 3 years total. During that time, 5 failures occur with repair times in days of 5, 10, 15, 8, and 12. What is the MTTR?",
-  answer("15 days"),
-  answer("5 days"),
-  answer("10 days", correct = TRUE),
-  answer("8 days"),
-  random_answer_order = TRUE
+```{r quiz3, echo=FALSE}
+quiz(caption = "Exercise: MTTR",
+  question("What does MTTR stand for?",
[…]
