+++ inst/tutorials/WeibullRlearnr/WeibullRlearnr.Rmd
@@ -95,25 +95,40 @@ Where $\beta$ is the shape parameter. In life data analysis $\beta$ has a physic
 
 And $\eta$ is the scale parameter or the quantile at 63.2% probability
 
-### Exercises
-
-```{r beta, echo=FALSE}
-question("Which beta below may indicate a wear out failure mode?",
-  answer("1"),
-  answer("0.5"),
-  answer("2", correct = TRUE),
-  answer("0"),
-  random_answer_order = TRUE
-)
-```
-
-```{r eta, echo=FALSE}
-question("What does the parameter eta represent?",
-  answer("shape parameter"),
-  answer("scale parameter", correct = TRUE),
-  answer("failure rate"),
-  answer("mean time between failures"),
-  random_answer_order = TRUE
+```{r quiz1, echo=FALSE}
+quiz(caption = "Quiz: The Weibull Parameters",
+  question("Which beta indicates an increasing failure rate?",
+    answer("1", message = "Beta of 1 indicates a constant failure rate."),
+    answer("0.5", message = "Beta less than 1 indicates a decreasing failure rate."),
+    answer("2", correct = TRUE, message = "Correct! Beta greater than 1 indicates an increasing failure rate."),
+    answer("0", message = "Beta cannot be 0."),
+    random_answer_order = TRUE, 
+    allow_retry = TRUE
+  ),
+  question("What does the parameter eta represent?",
+    answer("shape parameter", message = "Eta is the scale parameter, not the shape parameter."),
+    answer("scale parameter", correct = TRUE, message = "Correct! Eta is the scale parameter."),
+    answer("failure rate", message = "Eta is not the failure rate."),
+    answer("mean time between failures", message = "Eta is not the mean time between failures."),
+    random_answer_order = TRUE,
+    allow_retry = TRUE
+  ),
+  question("What is the cumulative probability when t = eta?",
+    answer("0%", message = "At t = eta, the cumulative probability is not 0%."),
+    answer("50%", message = "At t = eta, the cumulative probability is not 50%."),
+    answer("63.2%", correct = TRUE, message = "Correct! At t = eta, the cumulative probability is 63.2%."),
+    answer("100%", message = "At t = eta, the cumulative probability is not 100%."),
+    random_answer_order = TRUE,
+    allow_retry = TRUE
+  ),
+  question("What is the cumulative probability when t = 0?",
+    answer("0%", correct = TRUE, message = "Correct! At t = 0, the cumulative probability is 0%."),
+    answer("50%", message = "At t = 0, the cumulative probability is not 50%."),
+    answer("63.2%", message = "At t = 0, the cumulative probability is not 63.2%."),
+    answer("100%", message = "At t = 0, the cumulative probability is not 100%."),
+    random_answer_order = TRUE,
+    allow_retry = TRUE
+  )
 )
 ```
 
@@ -174,33 +189,50 @@ fit<-MLEw2p(failures, bounds=TRUE, show=TRUE)
 
 ### The Weibull Probability Plot
 
-The Weibull Probability Plot shows the time (or equivalent) on the horizontal axis and the unreliability on the vertical axis. In this case, unreliability is analogous to the cumulative probability of failure. We know from the cumulative distribution function that the opposite of the failure probability is the survival probability or reliability. So an unreliability of 80% at a given time is equivalent to a reliability of 20% (100% - 80% = 20%).
[…]
