### 2025-09-18 937920c
Update examples
