+++ inst/tutorials/TestRlearnr/TestRlearnr.Rmd
@@ -41,7 +41,7 @@ By the end of this tutorial, learners will be able to:
 
 ## The Duane Model
 
-**The Duane Model** was developed by J. T. Duane in 1964 and provides a simple and graphical way to observe and analyze whether failure rates are improving as changes are made to a product or system. 
+**The Duane model** provides a simple and graphical way to observe and analyze whether failure rates are improving as changes are made to a product or system. 
 
 Specifically, the Duane Model is a log-log plot of the cumulative Mean Time Between Failures (MTBF) vs cumulative time.
 
@@ -97,32 +97,11 @@ Where:
   - If \( \beta = 1 \), the system reliability is constant (no improvement or degradation).
   - If \( \beta < 1 \), the system reliability is deteriorating (the failure rate increases over time).
 
-### Exercises
-
-```{r duane1, echo=FALSE}
-question("What does a Duane Plot with a slope of greater than 1 indicate?",
-  answer("A) The system's reliability is improving", correct = TRUE),
-  answer("B) The system's reliability is deteriorating"),
-  answer("C) The system has no change in reliability"),
-  random_answer_order = TRUE
-)
-```
-
-```{r duane2, echo=FALSE}
-question("Which beta value may indicate an infant mortality failure mode?",
-  answer("1"),
-  answer("0.5"),
-  answer("2"),
-  answer("Unable to be determined in this case", correct = TRUE),
-  random_answer_order = TRUE
-)
-```
-
 ## ReliaGrowR
 
 ### Getting Started with ReliaGrowR
 
-For this tutorial, we will use **ReliaGrowR**: a R package for Reliability Growth Analysis.
+For this tutorial, we will use **ReliaGrowR**: an R package for Reliability Growth Analysis.
 
 First, check if ReliaGrowR is installed in R and install if not.
 
@@ -130,38 +109,63 @@ First, check if ReliaGrowR is installed in R and install if not.
 if(!(require(ReliaGrowR))){install.packages('ReliaGrowR')}
 ```
 
-Next, set up some cumulative time and failure data. Then use the `duane_plot` function to fit a Duane Model to the data. 
+Next, set up some cumulative time and failure data. Then use the `duane` function to fit a Duane Model to the data and the `plot` function to plot the results:
 
 ```{r echo=TRUE}
 times <- c(100, 200, 300, 400, 500)
 failures <- c(1, 2, 1, 3, 2)
-fit <- duane_plot(times, failures, line_col = "red")
+fit <- duane(times, failures)
+plot(fit, main = "Duane Model Example", xlab = "Cumulative Time", ylab = "Cumulative MTBF")
 ```
 
-### Exercises
+The plot shows a Duane Model fit to the data. The slope of the line indicates whether the system is improving, stable, or worsening.
 
-```{r duane3, echo=FALSE}
-question("In this example, what is the approximate cumulative MTBF at time 1000?",
-  answer("200"),
-  answer("100"),
-  answer("120"),
-  answer("150", correct = TRUE),
-  random_answer_order = TRUE
-)
-```
-
-```{r duane4, echo=FALSE}
-question("How would you describe the reliability of this system?",
-  answer("The system reliability is improving", correct = TRUE),
-  answer("The system reliability is worsening"),
-  answer("The system reliability is Stable"),
-  random_answer_order = TRUE
+```{r quiz1, echo=FALSE}
+quiz(caption = "Quiz: Duane Model",
+  question("What does a Duane Plot with a slope greater than 1 indicate?",
+    answer("The system's reliability is improving", correct = TRUE, 
+           message = "A slope greater than 1 indicates improvement."),
+    answer("The system's reliability is deteriorating", 
+           message = "A slope less than 1 indicates deterioration."),
+    answer("The system has no change in reliability", 
[…]
