+++ inst/paper/paper.md
@@ -40,7 +40,7 @@ Currently, three primary learning modules are available. These modules are indep
 and can be completed in any order. They are designed to be plug-and-play, but users 
 can modify them by forking the repository.
 
-`WeibullR.learnr()` provides an interactive introduction to Life Data Analysis. 
+`lda()` provides an interactive introduction to Life Data Analysis. 
 The learning objectives include basic Weibull analysis, censoring techniques such 
 as right and interval censored data, different types of Weibull models including 
 the 2P Weibull, 3P Weibull, and Weibayes model, parameter estimation methods Median
@@ -48,13 +48,13 @@ Rank Regression (MRR) and Maximum Likelihood Estimation (MLE), and data visualiz
 methods such as Probability Plots and Contour Plots. The estimated duration for 
 this module is approximately 2 hours. 
 
-`RAMR.learnr()` is a quick reference for common Reliability, Availability, and Maintainability
+`ram()` is a quick reference for common Reliability, Availability, and Maintainability
 (RAM) concepts. The learning objectives include the basic concepts and application 
 of Reliability, Availability, Mean Time to Repair (MTTR), Mean Time to Failure (MTTF), 
 Mean Time Between Failures (MTBF), Failure Rate, Probability of Failure, and $B_n$ 
 or $L_n$ life. The estimated duration of this module is about 1 hour.
 
-`TestR.learnr()` provides an interactive introduction to Reliability Testing. The 
+`rt()` provides an interactive introduction to Reliability Testing. The 
 learning objectives include defining key reliability growth concepts, including 
 Crow-AMSAA and Duane models, fitting a reliability growth model to real-world data
 using R, interpreting reliability growth plots and identifying trends, applying 
@@ -66,8 +66,10 @@ key patterns and data trends, and utilizing Arrhenius and Power Law models to ev
 the impact of stress factors on product reliability. The estimated duration of this 
 module is about 2 hours.
 
-The modules can also be accessed in a browser at [WeibullR.learnr](https://paulgovan.shinyapps.io/weibullrlearnr/),
-[RAMR.learnr](https://paulgovan.shinyapps.io/ramrlearnr/), and [TestR.learnr](https://govan.shinyapps.io/TestRlearnr/).
+The modules can also be accessed in a browser at 
+[https://paulgovan.shinyapps.io/RAMAnalysis/](https://paulgovan.shinyapps.io/RAMAnalysis/), 
+[https://paulgovan.shinyapps.io/LifeDataAnalysis/](https://paulgovan.shinyapps.io/LifeDataAnalysis/), 
+and [https://paulgovan.shinyapps.io/ReliabilityTesting/](https://paulgovan.shinyapps.io/ReliabilityTesting/).
 
 ![](https://github.com/paulgovan/WeibullR.learnr/blob/master/inst/paper/WeibullRlearnr.png?raw=true)<!-- -->
 
@@ -78,7 +80,6 @@ make it easy to apply the concepts covered in this module.
 * `avail()` - availability function
 * `mttf()` - mean time to failure
 * `mtbf()` - mean time between failure
-* `serv()` - serviceability factor
 * `fr()` - failure rate
 
 The project documentation includes installation instructions for `WeibullR.learnr`
