+++ man/RAMR.learnr.Rd
@@ -1,22 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RAMR.learnr.R
-\name{RAMR.learnr}
-\alias{RAMR.learnr}
-\title{Reliability, Availability, and Maintainability}
-\usage{
-RAMR.learnr()
-}
-\value{
-This function does not return a value.
-}
-\description{
-RAMR.learnr is an interactive introduction to RAM analysis.
-}
-\examples{
-if (interactive()) {
-  RAMR.learnr()
-}
-}
-\seealso{
-\url{https://paulgovan.github.io/WeibullR.learnr/}
-}
