+++ NEWS.md
@@ -1,4 +1,13 @@
-# WeibullR.learnr 0.2.1
+# ReliaLearnR 0.3 (Formerly WeibullR.learnr)
+
+## Breaking Changes
+* The package has been renamed from `WeibullR.learnr` to `ReliaLearnR` to better 
+reflect its broader focus on reliability engineering topics beyond just Weibull analysis.
+* All functions and tutorials have been updated to use the new package name `ReliaLearnR`.
+* The `WeibullR.learnr()` function has been renamed to `lda()` to simplify its usage. 
+* The `TestR.learnr()` function has been renamed to `rt()` to simplify its usage.
+* The `RAMR.learnr()` function has been renamed to `ram()` to simplify its usage.
+* All references to `WeibullR.learnr` in documentation and vignettes have been updated to `ReliaLearnR`.
 
 ## Minor Improvements and bug fixes
 
