+++ man/TestR.learnr.Rd
@@ -1,22 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/TestR.learnr.R
-\name{TestR.learnr}
-\alias{TestR.learnr}
-\title{An Interactive Introduction to Reliability Testing}
-\usage{
-TestR.learnr()
-}
-\value{
-This function does not return a value.
-}
-\description{
-TestR.learnr is an interactive introduction to Reliability Testing.
-}
-\examples{
-if (interactive()) {
-  TestR.learnr()
-}
-}
-\seealso{
-\url{https://paulgovan.github.io/WeibullR.learnr/}
-}

[…]
