+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.3
 pkgdown_sha: ~
 articles: {}
-last_built: 2025-09-20T23:12Z
+last_built: 2025-09-21T00:45Z
 urls:
   reference: https://paulgovan.github.io/ReliaLearnR/reference
   article: https://paulgovan.github.io/ReliaLearnR/articles
