+++ inst/paper/paper.bib
@@ -35,31 +35,48 @@
   doi = {10.32614/CRAN.package.WeibullR}
 }
 
-@Manual{WeibullRlearnr,
-  title = {WeibullR.learnr: An Interactive Introduction to Life Data Analysis},
+@Manual{WeibullRalt,
+    title = {WeibullR.ALT: Accelerated Life Testing Using 'WeibullR'},
+    author = {David Silkworth},
+    year = {2022},
+    note = {R package version 0.7.2},
+    url = {https://CRAN.R-project.org/package=WeibullR.ALT},
+  }
+
+@Manual{ReliaLearnR,
+  title = {ReliaLearnR: Learning Modules for Reliability Analysis},
   author = {Paul Govan},
-  year = {2023},
-  note = {R package version 0.1.1},
-  url = {https://CRAN.R-project.org/package=WeibullR.learnr},
-  doi = {10.32614/CRAN.package.WeibullR.learnr}
+  year = {2025},
+  note = {R package version 0.2},
+  url = {https://CRAN.R-project.org/package=ReliaLearnR},
+  doi = {10.32614/CRAN.package.ReliaLearnR}
 }
 
-@Manual{WeibullRplotly,
-  title = {WeibullR.plotly: Interactive Weibull Probability Plots with 'WeibullR'},
+@Manual{ReliaGrowR,
+    title = {ReliaGrowR: Reliability Growth Analysis},
+    author = {Paul Govan},
+    year = {2024},
+    note = {R package version 0.1},
+    url = {https://cran.r-project.org/package=ReliaGrowR},
+    doi = {10.32614/CRAN.package.ReliaGrowR},
+  }
+
+@Manual{ReliaPlotR,
+  title = {ReliaPlotR: Interactive Weibull Probability Plots},
   author = {Paul Govan},
   year = {2023},
   note = {R package version 0.1.5},
-  url = {https://CRAN.R-project.org/package=WeibullR.plotly},
-  doi = {10.32614/CRAN.package.WeibullR.plotly}
+  url = {https://CRAN.R-project.org/package=ReliaPlotR},
+  doi = {10.32614/CRAN.package.ReliaPlotR}
 }
 
-@Manual{WeibullRshiny,
-    title = {WeibullR.shiny: A 'Shiny' App for Weibull Analysis from 'WeibullR'},
+@Manual{ReliaShinR,
+    title = {ReliaShinR: A 'Shiny' App for Reliability Analysis},
     author = {Paul Govan},
     year = {2023},
     note = {R package version 0.1.0},
-    url = {https://CRAN.R-project.org/package=WeibullR.shiny},
-    doi = {10.32614/CRAN.package.WeibullR.shiny}
+    url = {https://CRAN.R-project.org/package=ReliaShinR},
+    doi = {10.32614/CRAN.package.ReliaShinR}
   }
 
 @book{Weibull,
