+++ inst/paper/paper.md
@@ -1,10 +1,14 @@
 ---
-title: 'WeibullR.learnr: An Interactive Introduction to Reliability Engineering'
+title: 'ReliaLearnR: Learning Modules for Reliability Analysis'
 tags:
   - R
   - life data analysis
   - Weibull analysis
   - reliability engineering
+  - availability
+  - maintainability
+  - reliability testing
+  - accelerated life testing
 authors:
   - name: Paul B. Govan
     orcid: 0000-0002-1821-8492
@@ -19,42 +23,53 @@ bibliography: paper.bib
 
 # Statement of Need
 
-Life data analysis is the study of how systems, from machines to people, perform 
-over time. Life data includes lifespans, failure rates, and repair or replacement
-timelines. While various learning resources exist, many rely on proprietary software
-that can be inaccessible to students and early-career professionals due to cost constraints.
+Reliability analysis is an important aspect of engineering that focuses on ensuring
+systems perform as expected over time. Key components of reliability analysis include
+life data analysis, reliability, availability, and maintainability (RAM) analysis,
+and reliability testing. These concepts are essential for engineers and professionals
+involved in product design, manufacturing, and maintenance. However, many learning resources
+for these topics rely on proprietary software, which can be expensive and inaccessible
+to students and early-career professionals. 
 
-`WeibullR.learnr` [@WeibullRlearnr] is an open-source collection of interactive 
-learning modules, exercises, and functions designed for introductory life data analysis. 
-The primary goal of this project is to introduce fundamental concepts while providing 
-an open-source alternative for analyzing life data. The target audience includes 
-beginner practitioners and university students.
+`ReliaLearnR` [@ReliaLearnR] addresses this gap by providing an open-source
+solution for learning reliability analysis using R [@R], a widely-used programming
+language for statistical computing and data analysis. The primary objective of this
+project is to introduce fundamental concepts of reliability analysis while providing
+an open-source alternative for analyzing reliability data. The target audience for
+this project includes beginner practitioners and university students interested
+in learning the fundamentals of reliability analysis.
 
 # Description
 
-`WeibullR.learnr` is written in R [@R] and utilizes `WeibullR` [@WeibullR], an R 
-package for Life Data Analysis based on the methodologies of Waloddi Weibull [@Weibull],
-and `learnr` [@learnr], a framework for building interactive learning modules in R.
+`ReliaLearnR` is writtn in R [@R] and utilizes `WeibullR` [@WeibullR], a 
+package for Life Data Analysis [@Weibull], `WeibullR.alt` [@WeibullRalt], a package
+for Accelerated Life Testing [@WeibullRalt], `ReliaGrowR` [@ReliaGrowR], a package
+for Reliability Growth Analysis [@ReliaGrowR], and `learnr` [@learnr], a framework for building 
+interactive learning modules.  
 
-Currently, three primary learning modules are available. These modules are independent
-and can be completed in any order. They are designed to be plug-and-play, but users 
-can modify them by forking the repository.
+The package includes three interactive learning modules, which can be run locally or
+hosted on a web server. Each module includes instructional content, code examples,
+and exercises to reinforce learning. Users can run the modules in any R environment 
+by calling the `lda()`, `ram()`, or `rt()` functions.
 
-`lda()` provides an interactive introduction to Life Data Analysis. 
-The learning objectives include basic Weibull analysis, censoring techniques such 
-as right and interval censored data, different types of Weibull models including 
-the 2P Weibull, 3P Weibull, and Weibayes model, parameter estimation methods Median
-Rank Regression (MRR) and Maximum Likelihood Estimation (MLE), and data visualization
[…]
