### 2025-09-21 paulgovan/ReliaLearnR#7 @paulgovan
To your point, I decided to rename the package from  `WeibullR.learnr` to `ReliaLearnR` to better
reflect its broader focus on reliability engineering topics beyond just Weibull analysis.
