+++ README.md
@@ -59,7 +59,7 @@ The modules can also be accessed in a browser at
 <https://paulgovan.shinyapps.io/LifeDataAnalysis/> and
 <https://paulgovan.shinyapps.io/ReliabilityTesting/>.
 
-![](https://github.com/paulgovan/WeibullR.learnr/blob/master/inst/paper/WeibullRlearnr.png?raw=true)<!-- -->
+![](https://github.com/paulgovan/ReliaLearnR/blob/master/inst/paper/ReliaLearnR.png?raw=true)<!-- -->
 
 Several helper functions for common RAM calculations are also included.
 These functions can be used independently of the learning modules:
