### 2025-09-22 #302 @paulgovan
Hi @lwjohnst86, @dbaidoo, and @AgnesDuah,

Thanks for your reviews — they led to some key improvements. The most notable is renaming the package from WeibullR.learnr to ReliaLearnR to reflect its broader reliability engineering focus. See [NEWS](https://github.com/paulgovan/ReliaLearnR/blob/master/NEWS.md) for details (not yet on CRAN).
I’ve responded to all issues: closed the straightforward ones and left open those needing more discussion.

FYI @stats-tgeorge
