+++ README.md
@@ -55,9 +55,10 @@ load the package and call the respective function:
   duration ~ 2 hours)
 
 The modules can also be accessed in a browser at
-<https://paulgovan.shinyapps.io/RAMAnalysis/>,
-<https://paulgovan.shinyapps.io/LifeDataAnalysis/> and
-<https://paulgovan.shinyapps.io/ReliabilityTesting/>.
+[paulgovan.shinyapps.io/RAMAnalysis/](https://paulgovan.shinyapps.io/RAMAnalysis/),
+[paulgovan.shinyapps.io/LifeDataAnalysis/](https://paulgovan.shinyapps.io/LifeDataAnalysis/)
+and
+[paulgovan.shinyapps.io/ReliabilityTesting/](https://paulgovan.shinyapps.io/ReliabilityTesting/).
 
 ![](https://github.com/paulgovan/ReliaLearnR/blob/master/inst/paper/ReliaLearnR.png?raw=true)<!-- -->
 
