### 2025-09-28 paulgovan/ReliaLearnR#32 @lwjohnst86
Not clear what exposed functions are for
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

The functions like `avail()`, `fr()`, etc are, more or less, simple calculations internally. If that's all they are, use the simple calculation in the learning material since it removes the chance to learn what is actually going on. Plus, learners using this learning module would have to use this package if they wanted to use these functions, which adds a dependency they have to keep in mind and use functions that aren't very "deep" (they are simple calculations). The specific issues I've identified with all these functions are:

- Are only used once (it seems only in `ram.Rmd`), so doesn't follow the "don't repeat yourself" principle as it isn't repeated again.
- Uses `print()` internally, which is very much an anti-pattern for making effective and robust functions.
- Adds errors via `stop()` that are already taken care of if using basic math on those checks (non-numbers and nulls would give an error in R already, these checks are not needed).
- Uses `sum()` internally for the arguments, which aren't needed as only scalars (single values) are passed into the function, and a sum of a single number is the number itself.p

Ultimately, these "helper" functions aren't helping the learner get practice and exposure to the concepts being taught, and take that learning opportunity away from them. Aside from that, the functions are extra code for simple math operations. (In software design, this is called "shallow design").

I suggest removing them and instead revise the learning material so that the learners can be exposed to the math of the problem domain.

If you really think they are valuable, pull them out into a separate function where you can build them more robustly and test them to ensure the match what you are aiming to build.
