### 2025-09-28 paulgovan/ReliaLearnR#34 @lwjohnst86
Describe the expectation pre-requisite knowledge of learners
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

Based on the learning objectives as well as the expected time to complete the lessons, I interpret this as learners already having quite a high level of pre-requisite knowledge. (e.g. a beginner to R would *not* be able to complete these tutorials in 2 hours and be able to learn anything from it in that time).

Could you describe some expectations you have of the learner in the README (or make a vignette with a "design" or "who are you" document)? Like, who they are, what their background knowledge is, what assumptions you make about them, etc. It would also help me a lot in reviewing, so that I know who the audience is. Because right now, I'm assume a relatively complete beginner, and that impacts how I review your tutorials.
