### 2025-09-28 paulgovan/ReliaLearnR#35 @lwjohnst86
Remove/modify estimated time to completion of lessons
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

After the revision of the learning objectives and related to #34, I think the time estimates for completing the tutorials *and actually learning and understanding anything* are very very optimistic. Of course, this also depends heavily on the existing knowledge the learners already have (part of the reason for #34).
