### 2025-09-28 da20738
refactor: :memo: move learning objectives to right after intro
