### 2025-09-28 paulgovan/ReliaLearnR#37 @lwjohnst86
Use full terminology when first introduced, then the abbreviation
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

In both `RAM` and `RT`, acronyms/abbreviations are used without writing them in full. Example: Crow-AMSAA is used in RT and MTTR, MTTF, and MTBF are used without describing what they are. In nearly all style guides, you always use/define the full term, followed by the abbreviation (as you did in LDA), then afterwards use the abbreviation.
