### 2025-09-28 paulgovan/ReliaLearnR#24 @lwjohnst86
This information should be in the README or in vignettes, not in a file that most learners/future users of the material can find it. The purpose of the paper file is mainly to direct teachers to the resource. See last paragraph of this section: https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain
