### 2025-09-28 paulgovan/ReliaLearnR#23 @lwjohnst86
There still needs to be a demonstration of how to actually complete the exercises. For example, in RAM, under section `## Reliability`, there is no demonstration of how to use the formula. The following exercise in the code chunk `quiz1` requires the learner know how to use the formulas. But that isn't a given assumption. How will they know how to use them without being shown how to use them?
