### 2025-09-28 paulgovan/ReliaLearnR#38 @lwjohnst86
`lda()`, `ram()`, and `rt()` don't work
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

Running these functions results in an error.
