### 2025-09-28 paulgovan/ReliaLearnR#22 @lwjohnst86
This is still needed even after the update. For example, in `lda.Rmd`, in the section `### Varying Beta`, there's no explanation about why we would want to or need to vary the beta, what that actually means when it changes, what the new change means, and why we should want to or need to manually vary the shape of slope in the plot. What information does that tell us as the learners? How does varying the slope of the plots work towards the learning objectives listed at the top?
