### 2025-09-28 paulgovan/ReliaLearnR#39 @lwjohnst86
Run styler on all code, including in the tutorials
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

Having a consistent style is critical to being able to effectively read and (try to) interpret code. For example, in `lda.Rmd` on line 186-187, the code is all cramped, with no spaces. Running styler will fix those issues.
