### 2025-09-28 paulgovan/ReliaLearnR#40 @lwjohnst86
Use simply `install.packages()`, without `require()`
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

You use `if(!(require(PACKAGENAME))){install.packages('PACKAGENAME')}` several times. This is redundant, as `install.packages()` already internally checks if the package is installed and if so, checks if it needs to be updated. So replace all those with `install.packages()`, as is.
