### 2025-09-28 paulgovan/ReliaLearnR#41 @lwjohnst86
Provide some narrative or example that ties the tutorials to the real-world
Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302

It helps a lot if learning material contains a strong narrative or basis in the real world. Could you use an example throughout each tutorial (doesn't have to be the same example in all three tutorials), that makes the content relatable?
