### 2025-09-28 fbe2a09
FIx tutorial launchers
