+++ R/rt.R
@@ -10,5 +10,5 @@
 #'   rt()
 #' }
 rt <- function() {
-  learnr::run_tutorial('rt', package = 'WeibullR.learnr')
+  learnr::run_tutorial('rt', package = 'ReliaLearnR')
 }
