+++ DESCRIPTION
@@ -20,5 +20,5 @@ Suggests:
     WeibullR.shiny
 License: Apache License (>= 2)
 Encoding: UTF-8
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
 VignetteBuilder: knitr
