### 2025-09-29 a1cd715
Fix tutorial launchers
