+++ DESCRIPTION
@@ -21,4 +21,3 @@ Suggests:
 License: Apache License (>= 2)
 Encoding: UTF-8
 RoxygenNote: 7.3.3
-VignetteBuilder: knitr
