### 2025-09-29 paulgovan/ReliaLearnR#40 @paulgovan
`install.packages()` will attempt to install a package even if the package is already installed, which is unnecessary overhead when the dependency is already installed and up to date. Using `require()` first allows a quick check: if the package is available, no further action; if not, then proceed to install.
