### 2025-09-29 26ed5ad
Remove time requirements
