+++ README.md
@@ -48,11 +48,9 @@ The package includes three interactive learning modules. To access them,
 load the package and call the respective function:
 
 - `ram()` - A quick reference for common Reliability, Availability, and
-  Maintainability concepts (estimated duration ~ 1 hour)
-- `lda()` - An introduction to Life Data Analysis (estimated duration ~2
-  hours)
-- `rt()`- An introduction to Reliability Testing (estimated duration ~ 2
-  hours)
+  Maintainability concepts
+- `lda()` - An introduction to Life Data Analysis
+- `rt()`- An introduction to Reliability Testing
 
 The modules can also be accessed in a browser at
 [paulgovan.shinyapps.io/RAMAnalysis/](https://paulgovan.shinyapps.io/RAMAnalysis/),
@@ -70,6 +68,21 @@ These functions can be used independently of the learning modules:
 - `mttf()` - mean time to failure
 - `mtbf()` - mean time between failure
 - `fr()` - failure rate
+
+## Motivation
+
+This project is based on a reliability program at a major technology
+company. The original program helped many early-career professionals
+learn the fundamentals of reliability engineering. However, the program
+was proprietary, limiting accessibility and leading to outdated learning
+materials as software evolved. By providing an open-source alternative,
+this project aims to reach a broader audience and foster a community of
+collaboration.
+
+# Acknowledgements
+
+The author acknowledges the creators of the original Reliability Program
+that inspired this project.
 
 ## Code of Conduct
 
