+++ README.Rmd
@@ -56,12 +56,10 @@ The package includes three interactive learning modules. To access
 them, load the package and call the respective function:
 
 -   `ram()` - A quick reference for common Reliability,
-    Availability, and Maintainability concepts (estimated duration \~ 1
-    hour)
+    Availability, and Maintainability concepts 
 -   `lda()` - An introduction to Life Data
-    Analysis (estimated duration \~2 hours)
+    Analysis
 -   `rt()`- An introduction to Reliability Testing
-    (estimated duration \~ 2 hours)
 
 The modules can also be accessed in a browser at
 [paulgovan.shinyapps.io/RAMAnalysis/](https://paulgovan.shinyapps.io/RAMAnalysis/),
@@ -78,6 +76,20 @@ functions can be used independently of the learning modules:
 -   `mttf()` - mean time to failure
 -   `mtbf()` - mean time between failure
 -   `fr()` - failure rate
+
+## Motivation
+
+This project is based on a reliability program at a major technology company. The
+original program helped many early-career professionals learn the fundamentals
+of reliability engineering. However, the program was proprietary, limiting accessibility
+and leading to outdated learning materials as software evolved. By providing an open-source 
+alternative, this project aims to reach a broader audience and foster a community 
+of collaboration. 
+
+# Acknowledgements
+
+The author acknowledges the creators of the original Reliability Program that inspired
+this project. 
 
 ## Code of Conduct
 
