### 2025-09-29 paulgovan/ReliaLearnR#34 @paulgovan
Added prerequisites in README (63c872b4b4bea2e132cbe5495b4e58856eca6309).
