+++ R/lda.R
@@ -11,5 +11,5 @@
 #'   lda()
 #' }
 lda <- function() {
-  learnr::run_tutorial('lda', package = 'ReliaLearnR')
+  learnr::run_tutorial("lda", package = "ReliaLearnR")
 }
