### 2025-09-29 paulgovan/ReliaLearnR#40 @lwjohnst86
Hm, that's true, I confused that with `pak()`, which is what I use much more often than `install.packages()`. Could you recommend installing with `pak()` instead? Since it is just a better package manager and installer than `install.packages()`. (e.g. it uses multiple threads when install, has better package dependency resolution algorithms, and also doesn't install if the package is the latest). If you use or expect your learners to use any tidyverse packages, like dplyr, or usethis package, than they will already have pak installed on their computers.
