### 2025-09-30 paulgovan/ReliaLearnR#32 @paulgovan
The helper functions are intended as quick calculation tools outside the tutorial. Even though they are simple, I’ve often seen these values miscalculated, so having a reliable function can be useful. In the tutorial, I introduce them only after the learner has already worked through the equations directly, so the math itself remains the primary focus.

I’d prefer not to remove them, but I agree they could be improved. I’ll plan to refine them further in future updates.
