+++ inst/tutorials/ram/ram.Rmd
@@ -43,9 +43,17 @@ That being said, all 3 terms are inter-related. A car that is not very reliable 
 Reliability is the probability that an item will not fail under defined conditions for a specified period.
 
 $$ Reliability = (1 - (Failed Time/Total Time))*100 $$
+Where Failed Time is the time that the item was not functioning, and Total Time is the total time that the item was in operation.
+
+Suppose an item ran for 5 years total and was failed for 10 of those days. The reliability of the item is: 
+
+$$ Reliability = (1 - (10/(5*365)))*100 = 94.5% $$
 Unreliability is the opposite of reliability, that is, the probability that an item will fail under defined conditions for a specified period. 
 
 $$ Unreliability = 1 - Reliability = (Failed Time/Total Time)*100 $$
+Take the previous example, the unreliability of the item is:
+
+$$ Unreliability = 1 - 0.945 = 0.055 = (10/(5*365))*100 = 5.5% $$
 
 ```{r quiz1, echo=FALSE}
 quiz(caption = "Exercise: Reliability",
@@ -73,8 +81,13 @@ quiz(caption = "Exercise: Reliability",
 Availability is the probability that an item will be available for service. Unavailable time includes failed time and scheduled maintenance.
 
 $$ Availability = (1 - (Unavailable Time/Total Time))*100 $$
+Where Unavailable Time is the time that the item was not available for service, and Total Time is the total time that the item was in operation.
 
 Unavailable time is different than standby time, where standby time is when the item is available for service, but is not being used. 
+
+Suppose an item ran for 5 years total, was failed for 10 days, and had scheduled maintenance for 15 days. The availability of the item is:
+
+$$ Availability = (1 - ((10+15)/(5*365)))*100 = 93.2% $$
 
 ```{r availability, echo=FALSE}
 quiz(caption = "Exercise: Availability",
@@ -102,6 +115,11 @@ quiz(caption = "Exercise: Availability",
 The Mean Time to Repair (MTTR) is a measure of maintainability and is the mean time to repair an item.
 
 $$ MTTR = \sum_{i=1}^n RepairTime_i/RepairCount $$
+Where $RepairTime_i$ is the time to repair for the *i*th repair, and $RepairCount$ is the total number of repairs.
+
+Suppose 100 items run for 5 years total. During that time, 5 failures occur with repair times in days of 5, 10, 15, 8, and 12. The MTTR is:
+
+$$ MTTR = (5+10+15+8+12)/5 = 10 days $$
 
 ```{r quiz3, echo=FALSE}
 quiz(caption = "Exercise: MTTR",
@@ -129,12 +147,20 @@ quiz(caption = "Exercise: MTTR",
 The Mean Time to Failure (MTTF) is a measure of reliability for *non-repairable* items, meaning that once the item fails, it is not repaired, but replaced. The MTTF is the ratio of total time by the number of failures for a given item.
 
 $$ MTTF = Total Time/FailureCount $$
+Where *Total Time* is the total time that the item was in operation, and *FailureCount* is the total number of failures.
 
+Suppose 100 items run for 5 years total. During that time, 5 failures occur. The MTTF is:
+
+$$ MTTF = (5*365)/5 = 365 years $$
 ## Mean Time Between Failures
 
 Mean Time Between Failures (MTBF) is a measure of reliability for *repairable systems*, meaning that once the item fails, it is repaired and put back into service. In this way, the item can fail multiple times. The MTBF is the ratio of total time by the number of failures for a given item.
 
[…]
