+++ R/rt.R
@@ -1,3 +1,4 @@
+
 #' Launch the Reliability Testing Tutorial
 #'
 #' This function launches the Reliability Testing tutorial
