+++ R/ram.R
@@ -1,3 +1,4 @@
+
 #' Launch the RAM Analysis Tutorial
 #'
 #' This function launches an interactive tutorial on Reliability, Availability,
