+++ man/fr.Rd
@@ -1,25 +1,25 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/lambda.R
+% Please edit documentation in R/helpers.R
 \name{fr}
 \alias{fr}
-\title{Failure Rate (lambda).}
+\title{Failure rate (lambda)}
 \usage{
 fr(failures, totalTime)
 }
 \arguments{
-\item{failures}{Failures. A numeric value representing the number of failures for a given period.}
+\item{failures}{Numeric scalar or numeric vector of failure counts (non-negative).}
 
-\item{totalTime}{Total Time. A numeric value representing the total time for a given period
-or a numeric vector of time periods}
+\item{totalTime}{Numeric scalar or numeric vector of total operating times.}
 }
 \value{
-The function returns a numeric value representing the failure rate (lambda) for a given period.
+Numeric scalar: failures per unit time (failures / totalTime).
 }
 \description{
-Failure Rate (lambda).
+Failure rate is the frequency with which an engineered system or component fails,
+expressed in failures per unit of time. This function computes failure rate given
+failure counts and total operating times.
 }
 \examples{
-fail <- 75
-total <- 5000
-fr(fail, total)
+fr(75, 5000)
+fr(c(10,5), c(1000,2000))
 }

[…]
