### 2025-09-30 02e1872
Delete test-launchers.R
