+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   helpers: helpers.html
-last_built: 2025-09-30T17:05Z
+last_built: 2025-09-30T17:14Z
 urls:
   reference: https://paulgovan.github.io/ReliaLearnR/reference
   article: https://paulgovan.github.io/ReliaLearnR/articles
