### 2025-10-01 paulgovan/ReliaLearnR#22 @paulgovan
Agreed. I went back through and found a few areas that could use better explanations and added more context. Let me know if you find anything else.
