+++ R/rt.R
@@ -2,7 +2,8 @@
 #' Launch the Reliability Testing Tutorial
 #'
 #' This function launches the Reliability Testing tutorial
-#' @import learnr ReliaGrowR WeibullR.ALT
+#' @importFrom learnr run_tutorial
+#' @import WeibullR WeibullR.ALT ReliaGrowR
 #' @export
 #' @seealso \url{https://paulgovan.github.io/ReliaLearnR/}
 #' @return This function does not return a value.
