+++ NAMESPACE
@@ -11,4 +11,4 @@ export(rt)
 import(ReliaGrowR)
 import(WeibullR)
 import(WeibullR.ALT)
-import(learnr)
+importFrom(learnr,run_tutorial)
