+++ tests/testthat/test-helpers.R
@@ -1,63 +1,38 @@
 
-test_that("rel() works for scalar and vector inputs and returns numeric scalar", {
-  expect_equal(rel(100, 1000), 0.9)
-  expect_equal(rel(c(10, 20), c(500, 600)), 1 - (sum(c(10,20)) / sum(c(500,600))))
-  expect_length(rel(100, 1000), 1)
-  expect_type(rel(100, 1000), "double")
+test_that("rel() and avail() handle boundary cases correctly", {
+  # outage = total -> reliability = 0
+  expect_equal(rel(500, 500), 0)
+  # unavail = 0 -> availability = 1
+  expect_equal(avail(0, 500), 1)
+  # tiny floating point overdrawn outage (clamped to 0)
+  expect_equal(rel(100.0001, 100), 0)
+  # tiny floating point negative (clamped to 1)
+  expect_equal(avail(1e-12, 100), 1)
 })
 
-test_that("rel() clamps values to [0,1] when outages exceed total", {
-  # outage > total -> clamped to 0
-  expect_equal(rel(600, 500), 0)
-  # If outage equals zero and total > 0 -> 1
-  expect_equal(rel(0, 500), 1)
+test_that("rel() and avail() error when totalTime sums to zero", {
+  expect_error(rel(1, 0), "sum\\(totalTime\\) is zero")
+  expect_error(avail(1, 0), "sum\\(totalTime\\) is zero")
+  # also with vectors summing to zero
+  expect_error(rel(c(1,2), c(0, -2)), "times must be non-negative")
 })
 
-test_that("avail() behaves like rel() (scalar, vector, clamp)", {
-  expect_equal(avail(100, 1000), 0.9)
-  expect_equal(avail(c(5,10), c(500,600)), 1 - (sum(c(5,10)) / sum(c(500,600))))
-  expect_equal(avail(600, 500), 0)   # clamped
-  expect_length(avail(10, 100), 1)
-  expect_type(avail(10, 100), "double")
+test_that("mtbf()/mttf() handle vector zero failures properly", {
+  expect_warning(mtbf(c(0,0), c(100,200)), "zero failures")
+  expect_true(is.infinite(suppressWarnings(mtbf(c(0,0), c(100,200)))))
+  expect_warning(mttf(c(0,0), c(50,150)), "zero failures")
+  expect_true(is.infinite(suppressWarnings(mttf(c(0,0), c(50,150)))))
 })
 
-test_that("mtbf() and mttf() compute correctly for normal cases", {
-  expect_equal(mtbf(5, 1000), 1000 / 5)
-  expect_equal(mttf(2, 500), 500 / 2)
-  expect_length(mtbf(5, 1000), 1)
-  expect_type(mttf(2, 500), "double")
+test_that("fr() errors when total time is zero", {
+  expect_error(fr(1, 0), "sum\\(totalTime\\) is zero")
+  expect_error(fr(c(1,2), c(0,0)), "sum\\(totalTime\\) is zero")
 })
 
-test_that("mtbf()/mttf() handle zero failures by returning Inf and emitting a warning", {
-  expect_warning(mtbf(0, 1000), "zero failures")
-  expect_true(is.infinite(suppressWarnings(mtbf(0, 1000))))
-  expect_warning(mttf(0, 2000), "zero failures")
-  expect_true(is.infinite(suppressWarnings(mttf(0, 2000))))
+test_that("all functions accept numeric vectors and collapse correctly", {
+  expect_equal(rel(c(10,10), c(50,50)), 1 - (20 / 100))
+  expect_equal(avail(c(1,1,1), c(10,10,10)), 1 - (3 / 30))
[…]
