### 2025-10-06 82c3f24
Update RAM examples
