+++ inst/tutorials/ram/ram.Rmd
@@ -28,7 +28,7 @@ By the end of this module, learners will be able to:
 
 ## What is RAM?
 
-The terms Reliability, Availability, and Maintainability may sound similar, but they actually represent different concepts.
+The terms Reliability, Availability, and Maintainability are related but distinct concepts.
 
 **Reliability**: The ability of an item to perform its intended function without failure over a specified period.
 
@@ -36,7 +36,7 @@ The terms Reliability, Availability, and Maintainability may sound similar, but 
 
 **Maintainability**: The ease and speed with which an item can be restored to operational status after a failure.
 
-That being said, all 3 terms are inter-related. A car that is not very reliable is probably not available very often. A car that is also difficult to maintain is probably available even less. 
+These concepts are interrelated in the sense that a highly reliable item is likely to be available more often, and an item that is easy to maintain can be restored to service quickly, thus improving availability.
 
 ## Reliability
 
@@ -45,19 +45,19 @@ Reliability is the probability that an item will not fail under defined conditio
 $$ Reliability = (1 - (Failed Time/Total Time))*100 $$
 Where Failed Time is the time that the item was not functioning, and Total Time is the total time that the item was in operation.
 
-Suppose an item ran for 5 years total and was failed for 10 of those days. The reliability of the item is: 
+Suppose a motor ran for 5 years total and was failed for 10 of those days. The reliability of the motor is: 
 
 $$ Reliability = (1 - (10/(5*365)))*100 = 94.5% $$
 Unreliability is the opposite of reliability, that is, the probability that an item will fail under defined conditions for a specified period. 
 
 $$ Unreliability = 1 - Reliability = (Failed Time/Total Time)*100 $$
-Take the previous example, the unreliability of the item is:
+Take the previous example, the unreliability of the motor is:
 
 $$ Unreliability = 1 - 0.945 = 0.055 = (10/(5*365))*100 = 5.5% $$
 
 ```{r quiz1, echo=FALSE}
 quiz(caption = "Exercise: Reliability",
-  question("An item has run for 3 years total and was failed for 5 of those days. What is the reliability of the item?",
+  question("A machine has run for 3 years total and was failed for 5 of those days. What is the reliability of the machine?",
     answer("90%", message = "Not quite, try again."),
     answer("0.5%", message = "0.5% is the unreliability, not the reliability."),
     answer("99.5%", correct = TRUE),
@@ -65,7 +65,7 @@ quiz(caption = "Exercise: Reliability",
     random_answer_order = TRUE,
     allow_retry = TRUE
   ),
-  question("What is the unreliability of the item?",
+  question("What is the unreliability of the machine?",
     answer("0.5%", correct = TRUE),
     answer("90%", message = "Not quite, try again."),
     answer("99.5%", message = "99.5% is the reliability, not the unreliability."),
@@ -85,7 +85,7 @@ Where Unavailable Time is the time that the item was not available for service, 
 
 Unavailable time is different than standby time, where standby time is when the item is available for service, but is not being used. 
 
-Suppose an item ran for 5 years total, was failed for 10 days, and had scheduled maintenance for 15 days. The availability of the item is:
+Suppose a motor ran for 5 years total, was failed for 10 days, and had scheduled maintenance for 15 days. The availability of the motor is:
 
 $$ Availability = (1 - ((10+15)/(5*365)))*100 = 93.2% $$
 
@@ -99,7 +99,7 @@ quiz(caption = "Exercise: Availability",
[…]
