### 2025-10-07 f8ba0eb
Update LDA examples
