+++ inst/tutorials/lda/lda.Rmd
@@ -13,7 +13,7 @@ library(learnr)
 
 ## Introduction
 
-Welcome to the tutorial on **Life Data Analysis**! In this tutorial, you will learn about the basics of life data analysis and how to apply it to various fields, including automotive, aerospace, and many others.
+Welcome to the tutorial on **Life Data Analysis**! In this tutorial, you will learn about the basics of life data analysis and how it applies to various fields, such as engineering, manufacturing, and many others. 
 
 ## Learning Objectives
 
@@ -27,11 +27,11 @@ By the end of this module, learners will be able to:
 
 ## What is Life Data Analysis?
 
-Life data analysis is the study of how systems, from machines to people, function over time. For example, in the automotive industry, life data helps explain how long cars will last before they break down. 
+Life data analysis is the study of how systems function over time. Life data helps explain how long a system will last, when it will likely fail, and how often it will need maintenance. For example, in the manufacturing industry, life data analysis can be used to understand the lifespan of machines and equipment, helping to optimize maintenance schedules and reduce downtime.
 
 ## Why is Life Data Analysis Important?
 
-Understanding life data analysis helps in predicting the lifespan and maintenance needs of various systems, thereby improving reliability and efficiency. By collecting and modeling life data, we can make predictions about how long cars will last or how long before they will need to be repaired or replaced. 
+Understanding life data analysis helps in predicting the lifespan and maintenance needs of various systems, thereby improving reliability and efficiency. By collecting and modeling life data, we can make predictions about how long machines will last or how long before they will need to be repaired or replaced. 
 
 ## Weibull Analysis
 
@@ -199,7 +199,9 @@ First, check if WeibullR is installed in R and install if not.
 if(!(require(WeibullR))){install.packages('WeibullR')}
 ```
 
-Next, let's run a basic example using WeibullR. First, create some failure data for 5 different machines that fail at time 30, 49, 82, 90, and 96 respectively. Then use the `MLEw2p` function to fit a Weibull to the data using the MLE method. 
+As an example, suppose a factory has 5 machines that fail at different times. The factory wants to understand the reliability of the machines over time and predict when future failures may occur. The factory collects the failure data and uses Weibull analysis to model the reliability of the machines. The failure times are 30, 49, 82, 90, and 96 respectively. 
+
+To run Weibull analysis in R, first create a vector with the failure data, then use the `MLEw2p` function to fit a Weibull to the data.
 
 ```{r echo=TRUE}
 failures<-c(30, 49, 82, 90, 96)
@@ -257,7 +259,9 @@ In life data analysis, data is often **censored**. Censoring occurs when the exa
 
 Right censored data includes **suspensions** or units that have operated for a period of time without failure. 
 
-Let's add right censored data to the previous example. In addition to the 5 machines that failed, add suspensions for 3 machines that did not fail (right censored) at times 100, 45, and 10 respectively. Then use the `MLEw2p` function to fit a Weibull to the data using the MLE method.
+Now suppose that the factory has 3 additional machines that did not fail during the analysis period. The machines were suspended (i.e. removed from service) at times 100, 45, and 10 respectively.
+
+To add suspensions to the previous example, create a vector with the suspension data and add it to the `MLEw2p` function.
 
 ```{r echo=TRUE}
[…]
