### 2025-10-07 e867b31
Update lda.Rmd
