### 2025-10-07 paulgovan/ReliaLearnR#47 @lwjohnst86
Fully remove tracking of `.Rproj.user/` folder
Not sure why, even though I made the PR #33, you still have this folder in the Git repo. Can you remove it fully and stop tracking it?

Connected to the review here: https://github.com/openjournals/jose-reviews/issues/302
