### 2025-10-07 paulgovan/ReliaLearnR#22 @lwjohnst86
There's still a lot needed. For example, in `lda.Rmd` in the "Getting started with WeibullR" section, what does `bounds=TRUE` or `show=TRUE` mean? This is just one example, but within this Rmd file alone, there are more examples like this where code is not explained.
