### 2025-10-07 #302 @lwjohnst86
@stats-tgeorge I've reviewed the repo a few times and made several issues and PRs. @paulgovan has done quite a lot to improve the material (great job! 🎉), but one of the key criteria of a JOSE acceptance is *can another teacher use this material easily*? At the current state, there's way too much unknown or confusing or not fully explained stuff that would allow another teacher (for example me) to actually take the repo and use it to teach others. So I'm not able to check those items off of the review checklist.

I also won't be able to continue reviewing this, as it has already taken up a lot of my time and, at least for the next two months, I will be very busy with work deadlines and tasks. @stats-tgeorge you can decide what to do with that.
