### 2025-10-08 4046541
Update examples in tutorials
