+++ inst/tutorials/rt/rt.Rmd
@@ -13,7 +13,7 @@ library(learnr)
 
 ## Introduction
 
-Welcome to the tutorial on **Reliability Testing**! In this tutorial, you will learn about different types of reliability tests and how to apply them. These models are commonly applied in industries like aerospace, automotive, and electronics, where improving product reliability over time is critical.
+Welcome to the tutorial on **Reliability Testing**! In this tutorial, we will explore key concepts and models used in reliability testing, including the Crow-AMSAA model, Duane model, and Accelerated Life Testing (ALT) methods such as the Arrhenius and Power Law models. These models help engineers and reliability professionals assess and improve the reliability of products and systems over time. 
 
 ## Learning Objectives
 
@@ -29,21 +29,25 @@ By the end of this tutorial, learners will be able to:
 
 ## What is Reliability Testing?
 
-Reliability testing is a process to ensure that a product or system performs its intended functions under specified conditions over a defined period. For example, in the automotive industry, we may want to design a car that runs without breaking down for 5 to 10 years. 
+Reliability testing is a process to ensure that a product or system performs its intended functions under specified conditions over a defined period. For example, in the manufacturing industry, we may want to design a machine that runs without breaking down for 5 to 10 years.
 
 ## Why is Reliability Testing important?
 
-Reliability testing helps organizations predict the future reliability of products, reduce failure rates, and lower costs associated with repairs and downtime. 
+Reliability testing is crucial for several reasons: 
+* It helps identify potential failure modes and weaknesses in a product or system.
+* It ensures that products meet quality and safety standards.
+* It enhances customer satisfaction by delivering reliable products.
+* It reduces costs associated with warranty claims, repairs, and downtime.
 
 ## Reliability Growth Analysis
 
-**Reliability Growth Analysis (RGA)** is a method to monitor and improve, or grow, the reliability of a product or system over time. Often the product is improved due to changes in the design throughout the testing period.
+**Reliability Growth Analysis (RGA)** is a method used to monitor and improve the reliability of a product or system over time. It involves analyzing failure data collected during testing phases to identify trends and patterns in reliability performance. 
 
 ## The Duane Model
 
-**The Duane model** provides a simple and graphical way to observe and analyze whether failure rates are improving as changes are made to a product or system. 
+**The Duane Model**, developed by Duane in 1964, is one of the earliest and most widely used models for reliability growth analysis. The model is particularly useful due to its simplicity and ease of interpretation. The Duane plot is a graphical representation that helps visualize whether failure rates are improving, stable, or worsening over time.
 
-Specifically, the Duane Model is a log-log plot of the cumulative Mean Time Between Failures (MTBF) vs cumulative time.
+Specifically, the Duane Model is a log-log plot of the cumulative Mean Time Between Failures (MTBF) vs cumulative time. 
 
 The slope of the line on the plot indicates the rate of reliability growth:
 
@@ -115,16 +119,21 @@ First, check if ReliaGrowR is installed in R and install if not.
 if(!(require(ReliaGrowR))){install.packages('ReliaGrowR')}
 ```
 
-Next, set up some cumulative time and failure data. Then use the `duane` function to fit a Duane Model to the data and the `plot` function to plot the results:
[…]
