### 2025-10-08 086e1fd
Add code explanation
