+++ inst/tutorials/lda/lda.Rmd
@@ -9,6 +9,7 @@ runtime: shiny_prerendered
 
 ```{r setup, include=FALSE}
 library(learnr)
+library(WeibullR)
 ```
 
 ## Introduction
@@ -196,12 +197,12 @@ For this tutorial, we will use WeibullR: an R package for Weibull analysis.
 First, check if WeibullR is installed in R and install if not.
 
 ```{r echo=TRUE, message=FALSE}
-if(!(require(WeibullR))){install.packages('WeibullR')}
+pak::pkg_install("WeibullR")
 ```
 
 As an example, suppose a factory has 5 machines that fail at different times. The factory wants to understand the reliability of the machines over time and predict when future failures may occur. The factory collects the failure data and uses Weibull analysis to model the reliability of the machines. The failure times are 30, 49, 82, 90, and 96 respectively. 
 
-To run Weibull analysis in R, first create a vector with the failure data, then use the `MLEw2p` function to fit a Weibull to the data.
+To run Weibull analysis in R, first create a vector with the failure data, then use the `MLEw2p` function with the failure data to fit a Weibull to the data. The argument `bounds=TRUE` adds confidence bounds to the fit, and `show=TRUE` displays the results.
 
 ```{r echo=TRUE}
 failures<-c(30, 49, 82, 90, 96)
@@ -288,7 +289,7 @@ A number of units did not fail during the inspection period. These units are con
 suspensions<-data.frame(time=63.48, event=0, qty=73)
 ```
 
-To run the interval censored example in WeibullR, use the `wblr` function with the `interval` argument to specify the inspection data. Then use the `wblr.fit` and `wblr.conf` functions to fit the model and add confidence bounds. Finally, use the `plot` function to create the probability plot.
+To run the interval censored example in WeibullR, use the `wblr` function with the suspension data and the `interval` argument to specify the inspection data. Then use the `wblr.fit` function with the `method.fit="mle"` argument to fit the model using maximum likelihood estimation. Next, use the `wblr.conf` function with the `method.conf="fm"` argument to add confidence bounds using the Fisher matrix method. The `col` and `lty` arguments specify the color and line type of the fit and confidence bounds respectively. Use the `plot` function to create the probability plot.
 
 ```{r echo=TRUE, warning=FALSE}
 obj1<-wblr(suspensions, interval = inspection_data)
@@ -303,11 +304,11 @@ Instead of points, the interval censored data is shown as black horizontal lines
 
 Life data is often compiled into groups (i.e. a group of machines failed or survived at a certain time). 
 
-Notice how the previous example contained groups in the `qty` column. Let's look at the inspection data again. 
+Notice how the previous example contained groups in the `qty` column. Let's look at the inspection data again. The knitr::kable function below displays the inspection data in a table format, and the head function shows the first few rows of the data.
 
 ``` {r echo=TRUE}
 knitr::kable(
-  head(inspection_data, topn=10)
+  head(inspection_data)
 )
 ```
 
@@ -360,7 +361,7 @@ quiz(caption = "Quiz: Interval Censored Data",
 
 WeibullR includes functions for both methods, such as the `MRRw2p`, `MRRw3p` functions for MRR and the `MLEw2p`, `MLEw3p` functions for MLE.
 
-To illustrate the differences between MRR and MLE, let's rerun the right censored example and compare the MRR and MLE methods using the `wblr` function.
[…]
