+++ inst/tutorials/rt/rt.Rmd
@@ -9,6 +9,8 @@ runtime: shiny_prerendered
 
 ```{r setup, include=FALSE}
 library(learnr)
+library(ReliaGrowR)
+library(WeibullR.ALT)
 ```
 
 ## Introduction
@@ -116,7 +118,7 @@ For this tutorial, we will use **ReliaGrowR**: an R package for Reliability Grow
 First, check if ReliaGrowR is installed in R and install if not.
 
 ```{r echo=TRUE, message=FALSE, warning=FALSE}
-if(!(require(ReliaGrowR))){install.packages('ReliaGrowR')}
+pak::pkg_install("ReliaGrowR")
 ```
 
 As an example, suppose a manufacturer wants to analyze the reliability growth of a new product. The product has been tested over a period of time, and the manufacturer has collected data on the cumulative operating time and the number of failures that occurred during that time. The cumulative time and failure data are as follows: 
@@ -126,7 +128,7 @@ times <- c(100, 200, 300, 400, 500)
 failures <- c(1, 2, 1, 3, 2)
 ```
 
-We can use the `duane` function to fit the model and the `plot` function to plot the results:
+We can use the `duane` function with the cumulative time and failure data to fit a Duane model. Then we can use the `plot` function to visualize the results. The `main`, `xlab`, and `ylab` parameters are used to customize the plot title and axis labels.
 
 ```{r echo=TRUE}
 fit <- duane(times, failures)
@@ -357,7 +359,6 @@ The plot shows a Piecewise NHPP Model fit to the data, with change points automa
 
 **Accelerated Life Testing (ALT)** is a method used to estimate the life of a product or system under normal operating conditions by subjecting it to elevated stress levels in a controlled environment. The goal is to identify potential failure modes and predict how long the product will last under normal conditions without having to wait for failures to occur naturally over a longer time period.
 
-
 ### Accelerated Life Testing: Stress vs Life
 
 Suppose you are testing a product that is expected to last a long time under normal conditions. Waiting for it to fail naturally could take months or even years. Instead, you can use ALT to accelerate the testing process by applying higher stress levels, such as increased temperature, voltage, or pressure, to induce failures more quickly.
@@ -409,12 +410,12 @@ For this tutorial, we will use **WeibullR.ALT**, an extension of the WeibullR pa
 First, check if WeibullR.ALT is installed in R and install if not.
 
 ```{r echo=TRUE, message=FALSE, warning=FALSE}
-if(!(require(WeibullR.ALT))){install.packages('WeibullR.ALT')}
+pak::pkg_install("WeibullR.ALT")
 ```
 
 To illustrate the Arrhenius Model, we will use the data set `MeekerData` that comes prepackaged with `WeibullR.ALT`. This data set contains time to failure data for carbon-film resistors tested at different temperature levels. The testing was conducted at 4 increasing temperature levels to accelerate the failure process. The temperature levels were 10, 40, 60, and 80 degrees Celsius.
 
-First, load the data set and view the first few samples.
+First, load the data set and use the `head` function to view the first few rows.
 
 ```{r}
 data <- MeekerData("table10")
@@ -423,7 +424,7 @@ head(data)
 
 The data set contains columns for the time to failure **time**, event type **event** (1 = failure, 0 = suspension), quantity of samples **qty** and temperature as the acceleration factor **TempC**. 
 
-Next, use the `alt.data` function to create a separate object for the data at each temperature level. 
[…]
