### 2025-10-08 #302 @paulgovan
Hi @lwjohnst86,

Thanks again for your feedback and for the time you’ve spent reviewing the repository.

I’ve now responded to all open issues, and these updates should address several of the concerns raised. I’ve also added references to relevant help pages and additional documentation to make the material more accessible for both learners and teachers.

Regarding the JOSE review criteria, it’s worth noting that the expectation is for instructors to have relevant expertise in the subject area. Someone with experience in reliability and a basic familiarity with R should be able to use these materials effectively to teach others. While the modules are designed to be approachable, they are not intended to be taught by someone without background in either reliability or R.
