### 2025-10-09 #302 @lwjohnst86
Thanks @paulgovan, yes I am aware that a teacher who would use this would also have the expertise in that area. My comment was not in relation to that, but made with the assumption the teacher had the expertise already. My comment was not about the content being confusing or not fully explained, but rather, *how would a teacher effectively know how to (re)use this material*, in a way that follows principles of learning and without too much effort in reusing the material effectively.
