### 2025-10-13 #302 @stats-tgeorge
Good morning, folks. I'm sorry for the delay. @lwjohnst86 I am very sorry to hear you cannot continue the review. I will contact our EIC and get her thoughts on moving forward. I'll get her opinion on your question, but this work falls within the aims and scopes of JOSE. @paulgovan we will get this done. It may or may not require another reviewer. Stay tuned.
