### 2025-10-13 #302 @lwjohnst86
@stats-tgeorge no worries. Just to be clear, what I meant with my comment/question is that I can't complete the checklist items (see my checklist) because they aren't addressed in the material/repository. Which (to me) is what tells me that it still doesn't fall under the criteria for JOSE. If I can't check the items off because the material don't address/fulfill the items, is that not necessary to be acceptable for submission?
