### 2025-10-13 #302 @stats-tgeorge
@lwjohnst86  I misunderstood your comment. This checklist of expectations, along with the aims and scopes/details outlined in the JOSE documentation, needs to be synthesized for future authors. I am very appreciative of your work so far on this!!

I agree that further changes are still needed. @lwjohnst86, can you suggest a path forward, rather than to say that the checklist isn't complete? I see issues in the repository that appear to have been considered by @paulgovan. Do these changes meet your expectations? If not, can you comment? For the other items, can you open more issues to make necessary updates?
