### 2025-10-13 #302 @lwjohnst86
I think the easiest way forward would be if @paulgovan reviewed the items I haven't checked off on the checklist (https://github.com/openjournals/jose-reviews/issues/302#issuecomment-3227180237) and try to revise the material to fit them. The reason I've done a several review passes is because initial suggested changes would greatly impact more detailed reviews (which would be a waste of time on my end and I assume on @paulgovan's).

For professional context, a big part of my job involves creating (and maintaining and running) teaching material that can (and is) reused by other teachers. I always have new teachers for each course/workshop I run, both to train more teachers and also to use them as "tests" to ensure my teaching material is reusable. Having that in mind, from a purely professional perspective, I would not be able to reuse @paulgovan's material in it's current state, though it of course can get to that state. So, as a professional, I cannot be able to check those items off. But those items do take a bit of work and effort to be able to achieve (I know from experience). And as I mentioned, the next ~2 months will be very busy for me and I won't be able to review until after that point.
