### 2025-10-13 #302 @paulgovan
@lwjohnst86, you mentioned one of the unchecked items is whether another teacher could easily use the material to teach, but there are several other unchecked items, even though they are addressed in the repository (e.g., the community guidelines are clearly defined). Could you clarify whether the unchecked items are areas you haven’t yet been able to review, or if they are areas you feel still need further work?

@stats-tgeorge, I’d be happy to open a few issues that map each unchecked review item to where it’s currently addressed in the repo. That way, you or a future reviewer can easily verify and decide whether any areas need further improvement.
