### 2025-10-13 #302 @lwjohnst86
@paulgovan the contributing guideline you have is the tidyverse guideline. It is a guide that they use for tidyverse packages, not to be used by others. It isn't specific to your repo, nor does many of it's content relate to your project. E.g. unit tests via `testthat` aren't needed in the context of a teaching repository. Likewise, you don't really use (or have that many) roxygen docs, since your project is a teaching material repo, not a typical R package (even though an R package is the method of delivery for your material). The remaining unchecked items are things I have decided to not make issues for at first because my earlier comments and suggestions would make some of those issues irrelevant (as I mentioned, they would have been a waste of time for me to review and you to respond to, since their relevance would quickly change based on the earlier reviews I made).
