### 2025-10-27 #302 @stats-tgeorge
Hello @paulgovan. Let's keep this moving forward. I agree that some updates are needed. I can have you make those changes, and then do my own checklist to make sure everything has been handled.

The modules are in the scope of JOSE.

_Doc_
Usage: Does the documentation explain how someone would adopt the module, and include examples of how to use it?

**There should be details on how to adopt and use the educational resource, for an educator, rather than package details**

Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support

**I agree that this needs to be expanded to how to contribute to the educational resource (your modules), not only the package components.**

_Pedagogy_
 Content scope and length: Is the content substantial for learning a given topic? Is the length of the module appropriate?

**There are several topics that need more detail to help someone learn from the provided content. Go through each step in the modules and add more explanations.**

Pedagogy: Does the module seem easy to follow? Does it observe guidance on cognitive load? (working memory limits of 7 +/- 2 chunks of information)

**With more details as suggested, this should be OK.**

Content quality: Is the writing of good quality, concise, engaging? Are the code components well crafted? Does the module seem complete?

**At a short glance (longer look later), looks good so far.**

Instructional design: Is the instructional design deliberate and apparent? For example, exploit worked-example effects; effective multi-media use; low extraneous cognitive load.

**Again, adding more details for prospective students/learners will go a long way.**

_JOSE Paper_
 Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?

**I don’t see this in the JOSE Paper**

Could someone else teach with this module, given the right expertise?

**Potentially after adding the suggested detail.**

Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?
References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

**There is some of this on the GH repo, but not in the paper.**
