### 2025-10-28 def027f
Add motivation to paper
