### 2025-10-28 paulgovan/ReliaLearnR#50 @paulgovan
Add "motivation" to the paper
This issue is meant to address the [JOSE](https://jose.theoj.org) review item:

Aa motivation section was added to the [paper](https://github.com/paulgovan/ReliaLearnR/blob/master/inst/paper/paper.md) to expand on how this project came about.

> # Motivation
>
> This project began as an effort to build upon a reliability program developed at
> a major technology company. The original program proved to provide a strong foundation,
> providing a structured learning opportunity that helped many early-career professionals
> understand and apply the fundamental concepts of reliability engineering.
> Over time, however, the proprietary nature of the program limited accessibility
> and adaptability.
>
> Recognizing the importance of keeping reliability learning both relevant and accessible,
> this project was initiated to create an open-source framework for teaching reliability
> analysis. By leveraging this framework, this project aims to reach a broader audience,
> encourage collaboration, and ensure that learning resources can evolve as needs and
> priorities change.
