### 2025-10-28 77d0dd2
Update motivation in paper
