### 2025-10-28 paulgovan/ReliaLearnR#51 @paulgovan
Add "design" to paper
This issue is meant to address the [JOSE](https://jose.theoj.org) review item:

A design section was added to the [paper](https://github.com/paulgovan/ReliaLearnR/blob/master/inst/paper/paper.md) to give more background on how these modules were used in the past and how they could be adopted for future use:

> # Design
>
> `ReliaLearnR` is written in R and utilizes `WeibullR` for Life Data
> Analysis, `WeibullR.alt` for Accelerated Life Testing,
> `ReliaGrowR` for Reliability Growth Analysis, and `learnr`,
> a framework for building interactive learning modules.
>
> The learning modules are designed to be interactive and engaging, with a
> focus on practical applications. Each module includes a mix of instructional content,
> code examples, and exercises to reinforce learning. The modules are self-paced,
> allowing learners to progress at their own speed.
>
> The original learning modules were provided in a series of workshops, where each
> workshop covered a specific module over a 1-2 hour period. These workshops were
> designed to be completed in a classroom setting with an instructor. The current version
> of the modules has been adapted for self-paced learning, but they can still be used
> in a classroom setting with an instructor.
>
> To adopt the modules for classroom use, instructors can either access them
> via the project website or install the package and use the functions directly.
> Instructors can also modify the modules to fit their specific needs, as the source code is
> available on the project repository.
