### 2025-10-28 1527e85
Add design to paper
