+++ README.Rmd
@@ -54,8 +54,9 @@ To do this when prompted during installation type 1 (All) to update.
 ## Recommended Background
 
 These modules are designed for engineering students and professionals interested
-in learning the fundamentals of reliability analysis. A basic understanding
-of R and statistics is recommended. For a complete beginners' guide to R, check out the resources at
+in learning the fundamentals of reliability analysis. While no prior experience is 
+required, a basic understanding of R and statistics is helpful. For a complete beginners' 
+guide to R, check out the resources at
 [https://education.rstudio.com/learn/beginner/](https://education.rstudio.com/learn/beginner/).
 
 ## Usage
@@ -76,7 +77,7 @@ The modules can also be accessed in a browser at
 
 ![](https://github.com/paulgovan/ReliaLearnR/blob/master/inst/paper/ReliaLearnR.png?raw=true)<!-- -->
 
-Several helper functions for common RAM calculations are also included. These
+The package also includes several helper functions for common RAM calculations. These
 functions can be used independently of the learning modules:
 
 -   `rel()` - reliability function
