### 2025-10-28 bfc1be6
Add design to README
