+++ _pkgdown.yml
@@ -1,6 +1,7 @@
 url: https://paulgovan.github.io/ReliaLearnR/
 template:
   bootstrap: 5
+  bootswatch: flatly
 development:
   mode: auto
 tutorials:
