+++ .github/CONTRIBUTING.md
@@ -1,20 +1,29 @@
 # Contributing to ReliaLearnR
 
-This outlines how to propose a change to ReliaLearnR.
-For a detailed discussion on contributing to this and other tidyverse packages, please see the [development contributing guide](https://rstd.io/tidy-contrib) and our [code review principles](https://code-review.tidyverse.org/).
+This guide outlines how to propose a change to ReliaLearnR.
 
 ## Fixing typos
 
 You can fix typos, spelling mistakes, or grammatical errors in the documentation directly using the GitHub web interface, as long as the changes are made in the _source_ file. 
-This generally means you'll need to edit [roxygen2 comments](https://roxygen2.r-lib.org/articles/roxygen2.html) in an `.R`, not a `.Rd` file. 
-You can find the `.R` file that generates the `.Rd` by reading the comment in the first line.
+This generally means you'll need to edit [rmarkdown](https://rmarkdown.rstudio.com) in an `.Rmd`, not an `.html` file. 
+You can find the `.Rmd` files for the tutorials in the '/inst/tutorials/<tutorial acronym>' sub directories. For example, the sub directory 
+for the life data analysis tutorial is located at '/inst/tutorials/lda/'.
 
 ## Bigger changes
 
 If you want to make a bigger change, it's a good idea to first file an issue and make sure someone from the team agrees that it’s needed. 
-If you’ve found a bug, please file an issue that illustrates the bug with a minimal 
-[reprex](https://www.tidyverse.org/help/#reprex) (this will also help you write a unit test, if needed).
-See our guide on [how to create a great issue](https://code-review.tidyverse.org/issues/) for more advice.
+If you’ve found a bug, please file an issue that illustrates the bug with a minimal example.
+
+## New tutorials
+
+This project is being actively developed, and recommendations for new tutorials are always welcome. While the aim is to cover a broad range of reliability topics,
+the focus is primarily on data-based reliability analysis methods. Some ideas for possible future tutorials include:
+
+* Repairable Systems Analysis
+* Parametric Survival Analysis
+* Reliability Block Diagrams
+
+To propose a new tutorial, or if you would like to contribute your own tutorial, please open an issue to discuss further. 
 
 ### Pull request process
 
@@ -28,17 +37,7 @@ See our guide on [how to create a great issue](https://code-review.tidyverse.org
     The title of your PR should briefly describe the change.
     The body of your PR should contain `Fixes #issue-number`.
 
-*  For user-facing changes, add a bullet to the top of `NEWS.md` (i.e. just below the first header). Follow the style described in <https://style.tidyverse.org/news.html>.
-
-### Code style
-
-*   New code should follow the tidyverse [style guide](https://style.tidyverse.org). 
-    You can use the [styler](https://CRAN.R-project.org/package=styler) package to apply these styles, but please don't restyle code that has nothing to do with your PR.  
-
-*  We use [roxygen2](https://cran.r-project.org/package=roxygen2), with [Markdown syntax](https://cran.r-project.org/web/packages/roxygen2/vignettes/rd-formatting.html), for documentation.  
-
-*  We use [testthat](https://cran.r-project.org/package=testthat) for unit tests. 
-   Contributions with test cases included are easier to accept.  
+*  For user-facing changes, add a bullet to the top of `NEWS.md` (i.e. just below the first header).
 
 ## Code of Conduct
 
