### 2025-10-28 paulgovan/ReliaLearnR#52 @paulgovan
Update contributing guidelines
This issue is meant to address the [JOSE](https://jose.theoj.org/) review item:

- [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support

Updated [contributing](https://github.com/paulgovan/ReliaLearnR?tab=contributing-ov-file) guidelines to add instructions for contributing to the tutorials/modules specifically.
