+++ README.Rmd
@@ -27,7 +27,7 @@ status](https://www.r-pkg.org/badges/version/ReliaLearnR)](https://CRAN.R-projec
 <!-- badges: end -->
 
 Welcome to **ReliaLearnR!** This package provides interactive learning
-modules for the fundamentals of reliability analysis. The modules are built
+modules on the fundamentals of reliability analysis. The modules are built
 using the `learnr` package and cover topics such as life data analysis,
 reliability testing, and reliability, availability, and maintainability
 (RAM) concepts. The package also includes helper functions for common
