+++ _pkgdown.yml
@@ -1,5 +1,6 @@
 url: https://paulgovan.github.io/ReliaLearnR/
 template:
+  light-switch: true
   bootstrap: 5
   bootswatch: flatly
 development:
