+++ inst/tutorials/ram/ram.Rmd
@@ -40,23 +40,26 @@ These concepts are interrelated in the sense that a highly reliable item is like
 
 ## Reliability
 
-Reliability is the probability that an item will not fail under defined conditions for a specified period.
+Reliability is the probability that an item will not fail under defined conditions for a specified period. Specifically, reliability is defined as:
 
 $$ Reliability = (1 - (Failed Time/Total Time))*100 $$
 Where Failed Time is the time that the item was not functioning, and Total Time is the total time that the item was in operation.
 
-Suppose a motor ran for 5 years total and was failed for 10 of those days. The reliability of the motor is: 
+For example, suppose a motor ran for 5 years total and was failed for 10 of those days. The reliability of the motor is: 
 
 $$ Reliability = (1 - (10/(5*365)))*100 = 94.5% $$
-Unreliability is the opposite of reliability, that is, the probability that an item will fail under defined conditions for a specified period. 
+So, the motor has a reliability of 94.5% over the 5-year period.
+
+Unreliability is the opposite of reliability, that is, the probability that an item _will_ fail under defined conditions for a specified period. Unreliability is defined as:
 
 $$ Unreliability = 1 - Reliability = (Failed Time/Total Time)*100 $$
 Take the previous example, the unreliability of the motor is:
 
 $$ Unreliability = 1 - 0.945 = 0.055 = (10/(5*365))*100 = 5.5% $$
+The motor has an unreliability of 5.5% over the 5-year period.
 
 ```{r quiz1, echo=FALSE}
-quiz(caption = "Exercise: Reliability",
+quiz(caption = "Exercise: Reliability", 
   question("A machine has run for 3 years total and was failed for 5 of those days. What is the reliability of the machine?",
     answer("90%", message = "Not quite, try again."),
     answer("0.5%", message = "0.5% is the unreliability, not the reliability."),
@@ -78,7 +81,7 @@ quiz(caption = "Exercise: Reliability",
 
 ## Availability
 
-Availability is the probability that an item will be available for service. Unavailable time includes failed time and scheduled maintenance.
+Availability is the probability that an item will be available for service. Unavailable time includes failed time and scheduled maintenance. Specifically, availability is defined as:
 
 $$ Availability = (1 - (Unavailable Time/Total Time))*100 $$
 Where Unavailable Time is the time that the item was not available for service, and Total Time is the total time that the item was in operation.
@@ -88,6 +91,7 @@ Unavailable time is different than standby time, where standby time is when the 
 Suppose a motor ran for 5 years total, was failed for 10 days, and had scheduled maintenance for 15 days. The availability of the motor is:
 
 $$ Availability = (1 - ((10+15)/(5*365)))*100 = 93.2% $$
+So the motor has an availability of 93.2% over the 5-year period.
 
 ```{r availability, echo=FALSE}
 quiz(caption = "Exercise: Availability",
@@ -112,7 +116,7 @@ quiz(caption = "Exercise: Availability",
 
 ## Mean Time to Repair
 
-The Mean Time to Repair (MTTR) is a measure of maintainability and is the mean time to repair an item.
+The Mean Time to Repair (MTTR) is a measure of maintainability and is defined as the average time required to repair a failed item. The MTTR is the ratio of total repair time by the number of repairs for a given item.
 
 $$ MTTR = \sum_{i=1}^n RepairTime_i/RepairCount $$
 Where $RepairTime_i$ is the time to repair for the *i*th repair, and $RepairCount$ is the total number of repairs.
@@ -144,24 +148,26 @@ quiz(caption = "Exercise: MTTR",
 
 ## Mean Time to Failure
 
[…]
