+++ inst/tutorials/rt/rt.Rmd
@@ -35,11 +35,12 @@ Reliability testing is a process to ensure that a product or system performs its
 
 ## Why is Reliability Testing important?
 
-Reliability testing is crucial for several reasons: 
-* It helps identify potential failure modes and weaknesses in a product or system.
-* It ensures that products meet quality and safety standards.
-* It enhances customer satisfaction by delivering reliable products.
-* It reduces costs associated with warranty claims, repairs, and downtime.
+Reliability testing is important for several reasons: 
+
+* Helps identify potential failure modes and weaknesses in a product or system.
+* Ensures that products meet quality and safety standards.
+* Enhances customer satisfaction by delivering reliable products.
+* Reduces costs associated with warranty claims, repairs, and downtime.
 
 ## Reliability Growth Analysis
 
@@ -49,7 +50,7 @@ Reliability testing is crucial for several reasons:
 
 **The Duane Model**, developed by Duane in 1964, is one of the earliest and most widely used models for reliability growth analysis. The model is particularly useful due to its simplicity and ease of interpretation. The Duane plot is a graphical representation that helps visualize whether failure rates are improving, stable, or worsening over time.
 
-Specifically, the Duane Model is a log-log plot of the cumulative Mean Time Between Failures (MTBF) vs cumulative time. 
+Specifically, the Duane Model is a log-log plot of the cumulative Mean Time Between Failures (MTBF) vs cumulative time, where the MTBF is calculated as the total operating time divided by the number of failures observed up to that time.
 
 The slope of the line on the plot indicates the rate of reliability growth:
 
@@ -359,7 +360,7 @@ The plot shows a Piecewise NHPP Model fit to the data, with change points automa
 
 **Accelerated Life Testing (ALT)** is a method used to estimate the life of a product or system under normal operating conditions by subjecting it to elevated stress levels in a controlled environment. The goal is to identify potential failure modes and predict how long the product will last under normal conditions without having to wait for failures to occur naturally over a longer time period.
 
-### Accelerated Life Testing: Stress vs Life
+### Acceleration Factor
 
 Suppose you are testing a product that is expected to last a long time under normal conditions. Waiting for it to fail naturally could take months or even years. Instead, you can use ALT to accelerate the testing process by applying higher stress levels, such as increased temperature, voltage, or pressure, to induce failures more quickly.
 
@@ -384,11 +385,11 @@ output$altPlot <- renderPlot({
 })
 ```
 
-In this plot, the blue curve represents the relationship between stress factor and expected life. The red dashed line indicates the expected life for the selected acceleration factor. As you increase the acceleration factor, the expected life decreases, demonstrating how elevated stress levels can significantly reduce the lifespan of a product.
+In this plot, the blue curve represents the relationship between the stress and the expected life. The red dashed line indicates the expected life for the selected acceleration factor. As you increase the acceleration factor, the expected life decreases, demonstrating how elevated stress levels can significantly reduce the lifespan of a product.
 
 ## The Arrhenius Model
 
-The **Arrhenius Model** is one of the most commonly used models in Accelerated Life Testing (ALT). It describes the relationship between temperature and the rate of chemical reactions, which can be directly related to the degradation or failure rate of materials. The model is based on the Arrhenius equation, which states that the rate of a chemical reaction increases exponentially with an increase in temperature.
[…]

[…]
