### 2025-11-02 #302 @paulgovan
Hi @stats-tgeorge — thanks again for the feedback. I’ve opened separate issues to track the remaining review items. When you have a chance, let me know if you agree with the changes or have more to add.

https://github.com/paulgovan/ReliaLearnR/issues/50
https://github.com/paulgovan/ReliaLearnR/issues/51
https://github.com/paulgovan/ReliaLearnR/issues/52
https://github.com/paulgovan/ReliaLearnR/issues/53
