### 2025-12-20 5887fbf
CRAN checks
