+++ NEWS.md
@@ -9,14 +9,33 @@ reflect its broader focus on reliability engineering topics beyond just Weibull 
 * The `RAMR.learnr()` function has been renamed to `ram()` to simplify its usage.
 * All references to `WeibullR.learnr` in documentation and vignettes have been updated to `ReliaLearnR`.
 
-## Minor Improvements and bug fixes
+## Other Improvements
+* Better (and more) examples and exercises in all tutorials. Improved comments and explanations.
+<https://github.com/paulgovan/ReliaLearnR/issues/21>, 
+<https://github.com/paulgovan/ReliaLearnR/issues/22>,
+<https://github.com/paulgovan/ReliaLearnR/issues/23>, 
+<https://github.com/paulgovan/ReliaLearnR/issues/25>, 
+<https://github.com/paulgovan/ReliaLearnR/issues/41>.
+* New RAM vignette <https://github.com/paulgovan/ReliaLearnR/issues/32>.
 
-* Minor Improvements and bug fixes
+## Minor Improvements and bug fixes
+* Improved package documentation 
+<https://github.com/paulgovan/ReliaLearnR/issues/20>, 
+<https://github.com/paulgovan/ReliaLearnR/issues/24>, 
+<https://github.com/paulgovan/ReliaLearnR/issues/26>, 
+<https://github.com/paulgovan/ReliaLearnR/issues/34>, 
+<https://github.com/paulgovan/ReliaLearnR/issues/35>,
+<https://github.com/paulgovan/ReliaLearnR/issues/50>,
+<https://github.com/paulgovan/ReliaLearnR/issues/51>,
+<https://github.com/paulgovan/ReliaLearnR/issues/52>,
+<https://github.com/paulgovan/ReliaLearnR/issues/50>.
+* Minor Improvements and bug fixes 
+<https://github.com/paulgovan/ReliaLearnR/issues/27>
 
 # WeibullR.learnr 0.2
 
 * Now with a new Reliability Testing tutorial called `TestR.learnr()` that covers Reliability Growth Analysis (RGA) and Accelerated Life Testing (ALT)
-* Other minor improvements and bug fixes
+* Other minor improvements and bug fixes <https://github.com/paulgovan/ReliaLearnR/issues/37>.
 
 # WeibullR.learnr 0.1.3
 
