+++ llms.txt
@@ -1,0 +1,143 @@
+# ReliaLearnR
+
+Welcome to **ReliaLearnR!** This package provides interactive learning
+modules on the fundamentals of reliability analysis. The modules are
+built using the `learnr` package and cover topics such as life data
+analysis, reliability testing, and reliability, availability, and
+maintainability (RAM) concepts. The package also includes helper
+functions for common RAM calculations.
+
+## Installation
+
+To install ReliaLearnR in R:
+
+``` r
+install.packages('ReliaLearnR')
+```
+
+To install the development version:
+
+``` r
+# install.packages("pak")
+pak::pak("paulgovan/ReliaLearnR")
+```
+
+Note: You may be prompted to update dependent packages before
+installing. To do so, type 1 (All) when prompted.
+
+## Recommended Background
+
+ReliaLearnR is designed for students and professionals who are
+interested in learning the fundamentals of reliability analysis. No
+prior experience is required, but a basic understanding of R and
+statistics is helpful. For a complete beginners’ guide to R, check out
+the resources at <https://education.rstudio.com/learn/beginner/>.
+
+## Usage
+
+The package includes three interactive learning modules. To launch the
+modules, load the package and call the respective function:
+
+- [`ram()`](https://paulgovan.github.io/ReliaLearnR/reference/ram.md) -
+  A quick reference for common Reliability, Availability, and
+  Maintainability (RAM) concepts
+- [`lda()`](https://paulgovan.github.io/ReliaLearnR/reference/lda.md) -
+  An introduction to Life Data Analysis
+- [`rt()`](https://paulgovan.github.io/ReliaLearnR/reference/rt.md)- An
+  introduction to Reliability Testing
+
+The modules can also be accessed in a browser at
+[paulgovan.shinyapps.io/RAMAnalysis/](https://paulgovan.shinyapps.io/RAMAnalysis/),
+[paulgovan.shinyapps.io/LifeDataAnalysis/](https://paulgovan.shinyapps.io/LifeDataAnalysis/)
+and
+[paulgovan.shinyapps.io/ReliabilityTesting/](https://paulgovan.shinyapps.io/ReliabilityTesting/).
+
+![](https://github.com/paulgovan/ReliaLearnR/blob/master/inst/paper/ReliaLearnR.png?raw=true)
+
+The package also includes several helper functions for common RAM
+calculations. These functions can be used independently of the learning
+modules:
+
+- [`rel()`](https://paulgovan.github.io/ReliaLearnR/reference/rel.md) -
+  reliability function
+- [`avail()`](https://paulgovan.github.io/ReliaLearnR/reference/avail.md) -
+  availability function
+- [`mttf()`](https://paulgovan.github.io/ReliaLearnR/reference/mttf.md) -
+  mean time to failure
+- [`mtbf()`](https://paulgovan.github.io/ReliaLearnR/reference/mtbf.md) -
+  mean time between failure
+- [`fr()`](https://paulgovan.github.io/ReliaLearnR/reference/fr.md) -
+  failure rate
+
+## Design
+
+The learning modules are designed to be interactive and engaging, with a
+focus on practical applications. Each module includes a mix of
+instructional content, code examples, and exercises to reinforce
+learning. The modules are self-paced, allowing learners to progress at
+their own speed.
+
+The original learning modules were provided in a series of workshops,
+where each workshop covered a specific module over a 1-2 hour period.
+These workshops were designed to be completed in a classroom setting
+with an instructor. The current version of the modules has been adapted
+for self-paced learning, but they can still be used in a classroom
+setting with an instructor.
+
+To adopt the modules for classroom use, instructors can either access
[…]
