+++ reference/lda.md
@@ -1,0 +1,25 @@
+# Launch the Life Data Analysis Tutorial
+
+This function launches an interactive tutorial for life data analysis.
+
+## Usage
+
+``` r
+lda()
+```
+
+## Value
+
+This function does not return a value.
+
+## See also
+
+<https://paulgovan.github.io/ReliaLearnR/>
+
+## Examples
+
+``` r
+if (interactive()) {
+  lda()
+}
+```
