+++ reference/ram.md
@@ -1,0 +1,26 @@
+# Launch the RAM Analysis Tutorial
+
+This function launches an interactive tutorial on Reliability,
+Availability, and Maintainability (RAM) analysis.
+
+## Usage
+
+``` r
+ram()
+```
+
+## Value
+
+This function does not return a value.
+
+## See also
+
+<https://paulgovan.github.io/ReliaLearnR/>
+
+## Examples
+
+``` r
+if (interactive()) {
+  ram()
+}
+```
