+++ reference/index.md
@@ -1,0 +1,20 @@
+# Package index
+
+## All functions
+
+- [`avail()`](https://paulgovan.github.io/ReliaLearnR/reference/avail.md)
+  : Availability (1 - unavailability / total)
+- [`fr()`](https://paulgovan.github.io/ReliaLearnR/reference/fr.md) :
+  Failure rate (lambda)
+- [`lda()`](https://paulgovan.github.io/ReliaLearnR/reference/lda.md) :
+  Launch the Life Data Analysis Tutorial
+- [`mtbf()`](https://paulgovan.github.io/ReliaLearnR/reference/mtbf.md)
+  : Mean Time Between Failures (MTBF) for repairable items.
+- [`mttf()`](https://paulgovan.github.io/ReliaLearnR/reference/mttf.md)
+  : Mean Time To Failure (MTTF)
+- [`ram()`](https://paulgovan.github.io/ReliaLearnR/reference/ram.md) :
+  Launch the RAM Analysis Tutorial
+- [`rel()`](https://paulgovan.github.io/ReliaLearnR/reference/rel.md) :
+  Reliability (1 - outage / total)
+- [`rt()`](https://paulgovan.github.io/ReliaLearnR/reference/rt.md) :
+  Launch the Reliability Testing Tutorial
