+++ tutorials/index.md
@@ -1,0 +1,8 @@
+# Tutorials
+
+- [Reliability, Availability, and Maintainability
+  (RAM)](https://paulgovan.github.io/ReliaLearnR/tutorials/ram.md)
+- [An Introduction to Life Data
+  Analysis](https://paulgovan.github.io/ReliaLearnR/tutorials/lda.md)
+- [An Introduction to Reliability
+  Testing](https://paulgovan.github.io/ReliaLearnR/tutorials/rt.md)

[…]
