+++ authors.md
@@ -10,17 +10,17 @@
 Source:
 [`inst/CITATION`](https://github.com/paulgovan/ReliaLearnR/blob/master/inst/CITATION)
 
-Govan P (2023). *WeibullR.learnr: An Interactive Introduction to Life
-Data Analysis*.
-[doi:10.32614/CRAN.package.WeibullR.learnr](https://doi.org/10.32614/CRAN.package.WeibullR.learnr),
+Govan P (2023). *ReliaLearnR: Learning Modules for Reliability
+Analysis*.
+[doi:10.32614/CRAN.package.ReliaLearnR](https://doi.org/10.32614/CRAN.package.ReliaLearnR),
 R package version 0.1.1,
-<https://CRAN.R-project.org/package=WeibullR.learnr>.
+<https://CRAN.R-project.org/package=ReliaLearnR>.
 
     @Manual{,
-      title = {WeibullR.learnr: An Interactive Introduction to Life Data Analysis},
+      title = {ReliaLearnR: Learning Modules for Reliability Analysis},
       author = {Paul Govan},
       year = {2023},
       note = {R package version 0.1.1},
-      url = {https://CRAN.R-project.org/package=WeibullR.learnr},
-      doi = {10.32614/CRAN.package.WeibullR.learnr},
+      url = {https://CRAN.R-project.org/package=ReliaLearnR},
+      doi = {10.32614/CRAN.package.ReliaLearnR},
     }
