+++ authors.md
@@ -8,19 +8,19 @@
 ## Citation
 
 Source:
-[`inst/CITATION`](https://github.com/paulgovan/ReliaLearnR/blob/master/inst/CITATION)
+[`inst/CITATION`](https://github.com/paulgovan/ReliaLearnR/blob/v0.3/inst/CITATION)
 
-Govan P (2023). *ReliaLearnR: Learning Modules for Reliability
-Analysis*.
-[doi:10.32614/CRAN.package.ReliaLearnR](https://doi.org/10.32614/CRAN.package.ReliaLearnR),
+Govan P (2023). *WeibullR.learnr: An Interactive Introduction to Life
+Data Analysis*.
+[doi:10.32614/CRAN.package.WeibullR.learnr](https://doi.org/10.32614/CRAN.package.WeibullR.learnr),
 R package version 0.1.1,
-<https://CRAN.R-project.org/package=ReliaLearnR>.
+<https://CRAN.R-project.org/package=WeibullR.learnr>.
 
     @Manual{,
-      title = {ReliaLearnR: Learning Modules for Reliability Analysis},
+      title = {WeibullR.learnr: An Interactive Introduction to Life Data Analysis},
       author = {Paul Govan},
       year = {2023},
       note = {R package version 0.1.1},
-      url = {https://CRAN.R-project.org/package=ReliaLearnR},
-      doi = {10.32614/CRAN.package.ReliaLearnR},
+      url = {https://CRAN.R-project.org/package=WeibullR.learnr},
+      doi = {10.32614/CRAN.package.WeibullR.learnr},
     }
