+++ index.md
@@ -1,7 +1,7 @@
 # ReliaLearnR
 
 Welcome to **ReliaLearnR!** This package provides interactive learning
-modules on the fundamentals of reliability analysis. The modules are
+modules on fundamental concepts in reliability analysis. The modules are
 built using the `learnr` package and cover topics such as life data
 analysis, reliability testing, and reliability, availability, and
 maintainability (RAM) concepts. The package also includes helper
