+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.2.0
 pkgdown_sha: ~
 articles:
   helpers: helpers.html
-last_built: 2026-01-06T15:09Z
+last_built: 2026-01-06T15:14Z
 urls:
   reference: https://paulgovan.github.io/ReliaLearnR/reference
   article: https://paulgovan.github.io/ReliaLearnR/articles
