### 2026-01-06 #302 @stats-tgeorge
@paulgovan I'm ready to push this forward. Let's get these small things resolved. Since it has been so long, it looks like you changed license. Please change to another OSI approved option. I'll start my final checklist too.  Please also look at the missing DOIs again.
