+++ DESCRIPTION
@@ -20,7 +20,7 @@ Suggests:
     ReliaShiny,
     rmarkdown,
     testthat (>= 3.0.0)
-License: CC BY 4.0
+License: Apache License (>= 2)
 Encoding: UTF-8
 RoxygenNote: 7.3.3
 VignetteBuilder: knitr
