### 2026-01-07 #302 @paulgovan
Thanks @stats-tgeorge! I changed the license back to Apache 2.0 as before.

DOIs:
- Added the DOI for WeibullR.ALT
- ReliaLearnR: the package was just re-indexed on CRAN (earlier today), so the DOI should be valid in the next day or two.

Should I also work on the author tasks listed above?
