### 2026-01-07 1e1a816
Update paper.bib
