+++ inst/paper/paper.bib
@@ -41,6 +41,7 @@
     year = {2022},
     note = {R package version 0.7.2},
     url = {https://CRAN.R-project.org/package=WeibullR.ALT},
+    doi = {10.32614/CRAN.package.WeibullR.ALT}
   }
 
 @Manual{ReliaLearnR,
