### 2026-01-07 #302 @stats-tgeorge
@paulgovan Thank you! Yes, please do.
