### 2026-01-07 #302 @stats-tgeorge
@paulgovan I also need an archive of your materials "When a submission is accepted, we ask that the authors create an archive (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or other) and post the archive DOI in the REVIEW issue."
