### 2026-01-07 f66f996
Update paper.md
