### 2026-01-07 #302 @paulgovan
> ## Additional Author Tasks After Review is Complete
>
> * [X]  Double check authors and affiliations (including ORCIDs)
> * [X]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSE paper. **Software v0.3.1**
> * [X]  Archive the release on Zenodo/figshare/etc and post the DOI here. [![DOI](https://zenodo.org/badge/644030886.svg)](https://doi.org/10.5281/zenodo.8010395)
> * [X]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSE paper.
> * [X]  Make sure that the license listed for the archive is the same as the software license.
