### 2026-01-08 #302 @stats-tgeorge
@paulgovan  the Zenodo archive name has to be the same as the JOSE paper title. Please update that.
