### 2026-01-08 #302 @lwjohnst86
@stats-tgeorge yea, because I created pull requests. The last I read of the JOSE reviewer policy was to make issues and pull requests. Maybe that's changed since I last read it?
