### 2026-01-08 #302 @lwjohnst86
For me, if I am already in the repo (locally) reviewing something, it's faster to make a PR for small things than to make an issue.
