### 2026-01-08 #302 @lwjohnst86
@stats-tgeorge ah, sorry I misunderstood your comment. @paulgovan it's fine to not include me as an author on the Zenodo record ☺️
