### 2026-01-08 #302 @stats-tgeorge
Oh I see! @paulgovan please update the Zenodo to only have you as an author. I am just making sure we don't have conflict of interest issues.
