+++ authors.md
@@ -8,7 +8,7 @@
 ## Citation
 
 Source:
-[`inst/CITATION`](https://github.com/paulgovan/ReliaLearnR/blob/v0.3.1/inst/CITATION)
+[`inst/CITATION`](https://github.com/paulgovan/ReliaLearnR/blob/master/inst/CITATION)
 
 Govan P (2023). *ReliaLearnR: Learning Modules for Reliability
 Analysis*.
