### 2026-01-08 #302 @paulgovan
@stats-tgeorge, I am learning Zenodo as I go, but I believe I made the necessary changes.

- Updated the title of the archive to match the title of the paper
- Removed @lwjohnst86 as an author
- I also made sure that the date of the archive matches the date listed in the paper

Here is the DOI for the entire archive:

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8010395.svg)](https://doi.org/10.5281/zenodo.8010395)

And here is the DOI for the specific release:

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.18176398.svg)](https://doi.org/10.5281/zenodo.18176398)
