### 2026-01-09 #302 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
