### 2026-01-09 #302 @stats-tgeorge
@openjournals/jose-eics I can't remember if it auto-notifies you, apologies if it does.
